### 2023-05-17 #5471 @kyleniemeyer
Hi @arend95, we'll use this pre-review issue to assign an editor and find reviewers - and also fix any initial issues with the submission.

Can you change the affiliation for the coauthor with multiple from `[1,2]` to `"1, 2"`?
