### 2023-05-17 #5471 @danielskatz
@arend95 - once the paper compiles, we can start looking for reviewers.

Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
