### 2023-05-18 #5478 @editorialbot
[REVIEW]: PyPO: a Python package for Physical Optics
**Submitting author:** @arend95 (<a href="http://orcid.org/0000-0002-0475-6134">Arend Moerman</a>)
**Repository:** https://github.com/PyPO-dev/PyPO
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
