### 2023-05-28 PyPO-dev/PyPO#128 @brandondube
conventional installation instructions, python packaging
This issue is part of my portion of the JOSS review, https://github.com/openjournals/joss-reviews/issues/5478

Ordinarily, python packages contain an installation heading on their main documentation page.  The PyPO docs require that I navigate from the index to "Related Pages" (a guess) and then the "General" page includes installation instructions.

Please place installation instructions on the index page of the documentation

--

The build instructions are highly non-normative for python, there is an ongoing multi-year period of confusion with python packaging, between poetry, pyproject.toml, setup.cfg/setup.py, and so on.  Regardless of all of that, the software should be installable using a one of the usual methods:

- `pip install .` or `pip install https://github.com/PyPO-dev/PyPO`
- `poetry add PyPO`

and so on.  This will require that you make a setup.py file, and/or setup.cfg+setup.py, or pyproject.tml file.

The user should not need to edit environment variables, and you should especially not prescribe path editing methods.  For example, I do not not even have `$PYTHONPATH` set; your instructions are incompatible with my ordinary, conda-based python environment

--

As-is, the instructions provide no feedback as to whether they worked or not; I ran `./Build.py -f` and then `-m`, which opened a new terminal for a brief moment with no feedback as to whether it worked or crashed.  Same for `-t`.
