### 2023-05-28 PyPO-dev/PyPO#129 @brandondube
documentation missing usage / how to launch
This issue is part of my portion of the JOSS review, https://github.com/openjournals/joss-reviews/issues/5478

The documentation contains three tutorials, which use the GUI.  There is not anywhere in the documentation (as best I can find) that shows how to launch the GUI, or even how to import PyPO when using the tool for scripting
