### 2023-05-28 PyPO-dev/PyPO#130 @brandondube
Unit tests are manual, not unit test?
This issue is part of my portion of the JOSS review, https://github.com/openjournals/joss-reviews/issues/5478

Browsing a random unit test, `test_Bessel.py`, as best I can tell the test does not do what it says:

```
# Script for testing the far-field calculations of PyPO by calculating the far-field of
# a uniformly illuminated disk, with varying wavelength, and calculating the HPBW.
# This is then compared to the theoretical value of a Bessel function.
```

The last line of the comment -- the comparison, seems to happen but not in a quantified way.  It looks like the assertion just tests whether the values are truth (not zero, NaN, inf, or -inf) - https://github.com/PyPO-dev/PyPO/blob/main/tests/test_Bessel.py#L33-L39

The plotting code seems to imply this is a test that a human runs, looks at a plot, and makes a decision on?  That is a manual validation test, not a unit test

I looked at a couple of other tests, e.g. fit gauss, and some of them are truly unit tests.  Please make all of the unit tests unit tests, and separate out the manual tests.

The tests also have a bad code smell,

```
from src.PyPO.System import System
```

You should never see this in python code, it should be `from PyPO.System import System`; the `src.` means this can only be done with cwd at the root of the repo, and implies an improperly installed package (related to #128)
