### 2023-05-28 PyPO-dev/PyPO#131 @brandondube
Accurate characterization of other work in JOSS paper
This issue is part of my portion of the JOSS review, https://github.com/openjournals/joss-reviews/issues/5478

The current draft of the JOSS paper discusses POPPY and prysm (caveat: latter is my software)

> Commercial software for GO and PO calculations and analysis, such as OpticStudio (Zemax) and GRASP (TICRA), has already been developed, but to our knowledge PyPO is the first free open-source Python package that simulates planar and quadric (off-axis) reflector geometries using both GO and PO. Moreover, PyPO does not employ approximations often employed by other software packages such as POPPy (Perrin & others, 2012) and Prysm (Dube, 2019).  Rather, PyPO directly solves the radiation integral, allowing for propagation between multiple (off-axis) reflector surfaces

At minimum, this paragraph should describe what approximation POPPY and prysm make that this software does not make.

If the "approximation" is a wave optics approach, the history for your repo begins October 27, 2022.  prysm's raytracer dates to December 2021 and can model all geometries supported by PyPO.

If the "approximation" is the solution to the Rayleigh-Sommerfeld integral in either the Fraunhofer or Fresnel regime, I think the above paragraph misrepresents things in a meaningful way.  POPPY and prysm use Fourier transforms to 'directly solve' the Rayleigh-Sommerfeld integral as it is written down in e.g. Born & Wolf or Goodman.  Those integrals are themselves approximations and assume that the cosine obliquity of the exit pupil is "small."  When it is not-small, one can include the cosine obliquity, which renders the integral exact once more.

If you want to make a claim that your software makes fewer approximations, I think you should include a demonstration that it is more accurate, as the purpose of making less approximations is surely higher accuracy.  The ray-trace result on tutorial 1, finding the focus, is significantly less accurate that one would expect for the focus of a parabola for an on-axis object at infinity.  That calculation should be of order ~ floating point epsilon.
