### 2023-05-28 PyPO-dev/PyPO#129 @brandondube
More generally, the documentation is extremely minimal.  Please create at least one tutorial showing how to use the software for scripting (a feature that is described in the paper).  Some documentation of the API (in the _ideal_ limit, a la Scipy/Numpy) should also be present, beyond auto-generated stubs showing the constructors and fields of structs, classes, etc.

A set of community guidelines and/or CONTRIBUTING.md are needed to complete the JOSS submission, as well.
