+++ MANIFEST.in
@@ -1,0 +1,8 @@
+include README.md 
+include LICENSE
+include CMakeLists.txt
+include requirements.txt
+recursive-include src *.py
+recursive-include src *.cpp 
+recursive-include src *.h 
+recursive-include src *.cu 
