+++ tests/test_SystemOps.py
@@ -7,7 +7,7 @@ import unittest
 import numpy as np
 from pathlib import Path
 
-from src.PyPO.System import System
+from PyPO.System import System
 
 class Test_SystemOps(unittest.TestCase):
     def setUp(self):
