+++ tests/test_SystemCopyAndRemove.py
@@ -1,7 +1,7 @@
 import unittest
 import numpy as np
-from src.PyPO.System import System as PyPOSystem
-from src.PyPO.Checks import InputReflError, InputRTError, InputPOError
+from PyPO.System import System as PyPOSystem
+from PyPO.Checks import InputReflError, InputRTError, InputPOError
 
 ##
 # @file
