+++ src/PyPO/System.py
@@ -1659,11 +1659,10 @@ class System(object):
     # @param show Show plot. Default is True.
     # @param save Save plot to savePath.
     # @param ret Return the Figure and Axis object. Only called by GUI. Default is False.
-    # @param extension Extension type of image to be saved.
     #
     # @returns fig Figure object.
     # @returns ax Axes object.
-    def plotBeamCut(self, name_field, comp, comp_cross=None, vmin=None, vmax=None, units='', name="", show=True, save=False, ret=False, extension="png"):
+    def plotBeamCut(self, name_field, comp, comp_cross=None, vmin=None, vmax=None, units='', name="", show=True, save=False, ret=False):
         E_cut, H_cut, E_strip, H_strip = self.calcBeamCuts(name_field, comp)
 
         #if comp_cross is not None:
@@ -1676,13 +1675,14 @@ class System(object):
         vmin = np.min([np.min(E_cut), np.min(H_cut)]) if vmin is None else vmin
         vmax = np.max([np.max(E_cut), np.max(H_cut)]) if vmax is None else vmax
         
-        fig, ax = plt.plotBeamCut(H_cut, E_cut, H_strip, E_strip, vmin, vmax, unitl)
+        fig, ax = plt.plotBeamCut(E_cut, H_cut, E_strip, H_strip, vmin, vmax, unitl)
 
         if ret:
             return fig, ax
 
         elif save:
-            pt.savefig(fname=self.savePath + f'{name}_EH_cut.{extension}', bbox_inches='tight', dpi=300)
+            pt.savefig(fname=self.savePath + '{}_EH_cut.jpg'.format(name),
+                        bbox_inches='tight', dpi=300)
             pt.close()
 
         elif show:
@@ -1915,7 +1915,6 @@ class System(object):
     # @param titleP Title of the phase plot. Default is "Phase".
     # @param unwrap_phase Unwrap the phase patter. Prevents annular structure in phase pattern. Default is False.
     # @param ret Return the Figure and Axis object. Only called by GUI. Default is False.
-    # @param extension Extension type of image to be saved.
     #
     # @see aperDict
     def plotBeam2D(self, name_obj, comp=None, contour=None, contour_comp=None,
@@ -1923,7 +1922,7 @@ class System(object):
                     save=False, interpolation=None, norm=True,
                     aperDict=None, mode='dB', project='xy',
                     units="", name="", titleA="Power", titleP="Phase",
-                    unwrap_phase=False, ret=False, extension="png"):
+                    unwrap_phase=False, ret=False):
 
         aperDict = {"plot":False} if aperDict is None else aperDict
 
@@ -1980,7 +1979,8 @@ class System(object):
             return fig, ax
 
         elif save:
-            pt.savefig(fname=self.savePath + f'{plotObject["name"]}_{name}.{extension}', bbox_inches='tight', dpi=300)
+            pt.savefig(fname=self.savePath + '{}_{}.jpg'.format(plotObject["name"], name),
+                        bbox_inches='tight', dpi=300)
             pt.close()
 
         elif show:
@@ -1999,9 +1999,8 @@ class System(object):
     # @param foc2 Plot focus 2. Default is False.
     # @param save Save the plot.
     # @param ret Return Figure and Axis. Only used in GUI.
-    # @param extension Extension type of image to be saved.
     def plot3D(self, name_surface, cmap=cm.cool,
-            norm=False, fine=2, show=True, foc1=False, foc2=False, save=False, ret=False, extension="png"):
+            norm=False, fine=2, show=True, foc1=False, foc2=False, save=False, ret=False):
         
         fig, ax = pt.subplots(figsize=(10,10), subplot_kw={"projection": "3d"})
         
@@ -2020,7 +2019,7 @@ class System(object):
[…]
