### 2023-06-02 dd4262d
pip sdist install more or less working now. Wheels not yet
