+++ src/PyPO/Plotter.py
@@ -397,40 +397,41 @@ def plotBeamCut(x_cut, y_cut, x_strip, y_strip, vmin, vmax, unit):
 # @param savePath Path to save plot to.
 # @param returns Whether to return figure object.
 # @param aspect Aspect ratio of plot.
-def plotRTframe(frame, project, savePath, returns, aspect):
+# @param unit Units of the axes for the plot.
+def plotRTframe(frame, project, savePath, returns, aspect, unit):
     fig, ax = pt.subplots(1,1, figsize=(5,5))
 
     idx_good = np.argwhere((frame.dx**2 + frame.dy**2 + frame.dz**2) > 0.8)
 
     if project == "xy":
-        ax.scatter(frame.x[idx_good], frame.y[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$x$ / mm")
-        ax.set_ylabel(r"$y$ / mm")
+        ax.scatter(frame.x[idx_good] * unit[1], frame.y[idx_good] * unit[1], color="black", s=10)
+        ax.set_xlabel(r"$x$ / {}".format(unit[0]))
+        ax.set_ylabel(r"$y$ / {}".format(unit[0]))
 
     elif project == "xz":
-        ax.scatter(frame.x[idx_good], frame.z[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$x$ / mm")
-        ax.set_ylabel(r"$z$ / mm")
+        ax.scatter(frame.x[idx_good] * unit[1], frame.z[idx_good] * unit[1], color="black", s=10)
+        ax.set_xlabel(r"$x$ / {}".format(unit[0]))
+        ax.set_ylabel(r"$z$ / {}".format(unit[0]))
     
     elif project == "yz":
-        ax.scatter(frame.y[idx_good], frame.z[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$y$ / mm")
-        ax.set_ylabel(r"$z$ / mm")
+        ax.scatter(frame.y[idx_good] * unit[1], frame.z[idx_good] * unit[1], color="black", s=10)
+        ax.set_xlabel(r"$y$ / {}".format(unit[0]))
+        ax.set_ylabel(r"$z$ / {}".format(unit[0]))
     
     elif project == "yx":
-        ax.scatter(frame.y[idx_good], frame.x[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$y$ / mm")
-        ax.set_ylabel(r"$x$ / mm")
+        ax.scatter(frame.y[idx_good] * unit[1], frame.x[idx_good] * unit[1], color="black", s=10)
+        ax.set_xlabel(r"$y$ / {}".format(unit[0]))
+        ax.set_ylabel(r"$x$ / {}".format(unit[0]))
 
     elif project == "zy":
-        ax.scatter(frame.z[idx_good], frame.y[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$z$ / mm")
-        ax.set_ylabel(r"$y$ / mm")
+        ax.scatter(frame.z[idx_good] * unit[1], frame.y[idx_good] * unit[1], color="black", s=10)
+        ax.set_xlabel(r"$z$ / {}".format(unit[0]))
+        ax.set_ylabel(r"$y$ / {}".format(unit[0]))
     
     elif project == "zx":
-        ax.scatter(frame.z[idx_good], frame.x[idx_good], color="black", s=10)
-        ax.set_xlabel(r"$z$ / mm")
-        ax.set_ylabel(r"$x$ / mm")
+        ax.scatter(frame.z[idx_good] * unit[1], frame.x[idx_good] * unit[1], color="black", s=10)
[…]
