+++ src/PyPO/System.py
@@ -38,9 +38,9 @@ logging.getLogger(__name__)
 #
 # This script contains the System class definition.
 class System(object):
-    customBeamPath = os.path.join(sysPath, "custom", "beam")
-    customReflPath = os.path.join(sysPath, "custom", "reflector")
-    savePathSystems = os.path.join(sysPath, "save", "systems")
+    customBeamPath = os.getcwd()
+    customReflPath = os.getcwd()
+    savePathSystems = os.getcwd()
 
 
     ##
@@ -81,7 +81,7 @@ class System(object):
         if not saveSystemsExist:
             os.makedirs(self.savePathSystems)
         
-        self.savePath = os.path.join(sysPath, "images")
+        self.savePath = os.getcwd() 
 
         existSave = os.path.isdir(self.savePath)
 
@@ -114,7 +114,12 @@ class System(object):
             del self.clog
     
     ##
-    # Obtain a reference to the custom logger used by system.
+    # @brief Obtain a reference to the custom logger used by system.
+    #
+    # This method can be called to obtain a reference to the logging object that PyPO uses internally.
+    # Can be convenient in cases one wants to log their own information in the layout of the PyPO logger.
+    #
+    # @ingroup public_api_sysio
     #
     # @returns clog Reference to system logger.
     def getSystemLogger(self):
@@ -145,10 +150,15 @@ class System(object):
         return s
 
     ##
-    # Set path to folder containing custom beams.
+    # @brief Set path to folder containing custom beam patterns.
     #
-    # @param path Path to custom beams. Beams should be stored in r<name_of_beam>.txt and i<name_of_beam>.txt format, containing real and imaginary parts respectively.
-    # @param append Whether path is relative to ./custom/beams/ or absolute.
+    # Set the path to the directory where PyPO looks for custom beam patterns.
+    # When a custom beam pattern is imported using the readCustomBeam() method, PyPO will look in the specified path for the beam patterns. 
+    #
+    # @ingroup public_api_sysio
+    #
+    # @param path Path to custom beams.
+    # @param append Whether path is relative to current working directory or absolute from home.
     def setCustomBeamPath(self, path, append=False):
         if append:
             self.customBeamPath = os.path.join(self.customBeamPath, path)
@@ -158,8 +168,12 @@ class System(object):
     ##
     # Set path to folder were to save output plots.
     #
+    # Set the path to the directory where PyPO saves output plots.
+    #
+    # @ingroup public_api_sysio
+    #
     # @param path Path to save directory.
-    # @param append Whether path is relative to ./images/ or absolute.
+    # @param append Whether path is relative to current working directory or absolute from home.
     def setSavePath(self, path, append=False):
         if append:
             self.savePath = os.path.join(self.savePath, path)
@@ -173,8 +187,12 @@ class System(object):
     ##
     # Set path to folder were to save systems.
     #
+    # Set the path to the directory where PyPO saves systems.
+    #
+    # @ingroup public_api_sysio
+    #
     # @param path Path to save directory.
-    # @param append Whether path is relative to ./save/systems/ or absolute.
+    # @param append Whether path is relative to current working directory or absolute from home.
     def setSavePathSystems(self, path, append=False):
         if append:
             self.savePathSystems = os.path.join(self.savePathSystems, path)
@@ -188,6 +206,11 @@ class System(object):
     ##
     # Set the verbosity of the logging from within the system.
     #
+    # Sometimes it is preferrable to not generate logging output to the console, for example when PyPO methods are called in for loops.
+    # This method allows for on/off switching of the logging unit so that the console is not flooded with logging messages.
+    #
+    # @ingroup public_api_sysio
+    #
     # @param verbose Whether to enable logging or not.
[…]
