+++ doxy/Tutorial.md
@@ -1,24 +1,0 @@
-<h1>Basic functionalities</h1>
-These tutorials are designed to showcase the features of PyPO.
-
-[Building an optical system and performing ray-traces](tutorials/Tutorial1.html)
-
-[More advanced optical systems and ray-tracing options](tutorials/Tutorial2.html)
-
-[Performing physical optics propagations](tutorials/Tutorial3.html)
-
-[Backward propagation](tutorials/Tutorial4.html)
-
-[PO efficiencies and metrics](tutorials/Tutorial5.html)
-
-[Importing custom beam patterns](tutorials/Tutorial6.html)
-
-<h1>GUI tutorials</h1>
-To provide a little help on how to use the GUI module, these tutorials can come in handy.
-The tutorials are designed to demonstrate basic knowledge on how to use `PyPO` through the GUI.
-
-[Tutorial 1](Gui/Tutorial_Gui_1.html)
-
-[Tutorial 2](Gui/Tutorial_Gui_2.html)
-
-[Tutorial 3](Gui/Tutorial_Gui_3.html)
