+++ MANIFEST.in
@@ -1,7 +1,6 @@
 include README.md 
 include LICENSE
 include CMakeLists.txt
-include requirements.txt
 recursive-include src *.py
 recursive-include src *.cpp 
 recursive-include src *.h 
