+++ CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.16.3)
 
 #set (CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS}")
 
-project(PyPO 
+project(PyPO-pkg 
     VERSION 1.0.0 
     DESCRIPTION "Python Physical Optics"
     LANGUAGES CXX C)
