### 2023-06-06 cfd8e6d
Packaging works. Working on better docs as well
