+++ src/PyPO/System.py
@@ -60,14 +60,75 @@ class System(object):
         Config.setContext(context)
         Config.setOverride(override)
 
-        # Internal dictionaries
+        ##
+        # The internal system dictionary.
+        #
+        # This dictionary holds the user-defined reflector/element dictionaries.
+        # The key of the entry is the name of the reflector/element.
+        # The reflector/element dictionary itself is stored as the item corresponding to the key.
+        #
+        # @ingroup public_api_internaldicts 
         self.system = {}
+        
+        ##
+        # The internal frames dictionary.
+        #
+        # This dictionary holds all ray-trace frames generated by `PyPO`.
+        # The frames are stored as key-value pairs, where the key is the name given to the frame.
+        # The item stored at the key is the frame itself.
+        #
+        # @ingroup public_api_internaldicts 
         self.frames = {}
+        
+        ##
+        # The internal PO fields dictionary.
+        #
+        # This dictionary holds all PO fields generated by `PyPO`.
+        # The fields are stored as key-value pairs, where the key is the name given to the PO field.
+        # The item stored at the key is the PO field itself.
+        #
+        # @ingroup public_api_internaldicts 
         self.fields = {}
+        
+        ##
+        # The internal PO currents dictionary.
+        #
+        # This dictionary holds all PO currents generated by `PyPO`.
+        # The currents are stored as key-value pairs, where the key is the name given to the PO current.
+        # The item stored at the key is the PO current itself.
+        #
+        # @ingroup public_api_internaldicts 
         self.currents = {}
+        
+        ##
+        # The internal PO scalarfields dictionary.
+        #
+        # This dictionary holds all PO scalarfields generated by `PyPO`.
+        # The scalarfields are stored as key-value pairs, where the key is the name given to the PO scalarfield.
+        # The item stored at the key is the PO scalarfield itself.
+        #
+        # @ingroup public_api_internaldicts 
         self.scalarfields = {}
 
+        ##
+        # The internal groups dictionary.
+        #
+        # This dictionary holds all defined groups.
+        # The key under which a group is stored is the name of that group.
+        #
+        # @ingroup public_api_internaldicts 
         self.groups = {}
+        
+        ##
+        # The internal associations dictionary.
+        #
+        # This dictionary contains an overview of Poynting vector ray-trace frames and the corresponding reflected PO field.
+        # The key in the dictionary is the name of the surface on which the frame and PO field are defined.
+        # In `PyPO` language, this is called an association, as in, the frame and PO field are associated on the same surface.
+        # This is important for doing hybrid propagations, where the field is propagated using the reflected Poynting vectors and reflected PO field.
+        # If a user attempts to do a hybrid propagation with a frame and PO field that are not associated, the `System` will log an error.
+        #
+        # @ingroup public_api_internaldicts
         self.assoc = {}
 
         self.EHcomplist = np.array(["Ex", "Ey", "Ez", "Hx", "Hy", "Hz"])
@@ -741,7 +802,8 @@ class System(object):
             check_elemSystem(name, self.system, self.clog, extern=True)
             self._checkBoundPO(name, InvertMat(self.system[name]["transf"]))
             self.system[name]["transf"] = self.copyObj(self.system[name]["snapshots"][snap_name])
-            self._checkBoundPO(name, self.system[name]["transf"]) self.clog.info(f"Reverted element {name} to snapshot {snap_name}.")
+            self._checkBoundPO(name, self.system[name]["transf"]) 
+            self.clog.info(f"Reverted element {name} to snapshot {snap_name}.")
         elif obj == "frame":
[…]
