### 2023-06-07 b24cd1d
Getting ready for putting it all together
