+++ tutorials/Gui/Tutorial_Gui_1.md
@@ -1,0 +1,108 @@
+# PyPO GUI Manual
+
+## Starting The GUI
+Before we can use the GUI, we have to start it up.
+The GUI can only ben used if `PyPO` is installed from source because the GUI source code is only included if directly obtained from Github.
+Also make sure that PySides2 is installed, otherwise the GUI will not launch.
+
+Navigate to the location of the `PyPO` root directory and run:
+```
+python GuiApp.py
+```
+
+which should launch the GUI.
+
+## Introduction
+
+In this tutorial we wil, step by step, build a simple reflector system and perform a ray-trace calculation trough it.
+<!--![Blank app](README_Resources/blankGui.png)-->
+
+The gui consists of 3 widgets:
+
+* Workspace (left)
+* Plot screen
+* Console
+
+For most interactions a form will show up next to the workspace.
+
+
+<figure><img src="README_Resources/blankGui.png" alt="A blank GUI window at startup" width="800px"/><figcaption>A blank GUI window at startup</figcaption></figure>
+
+
+
+## Creating Reflectors
+From the elements menu select *Add Reflector* > *Quadric surface*. 
+
+A form will show up with one dropdown selector. By selecting options, more options will show up. By filling in the parameters and clicking *add*, a reflector has been added to the system. If successful, a widget will show up in the workspace as such:
+
+<!--![System with one paraboloid reflector](README_Resources/pri_form.png)-->
+
+<figure><img src="README_Resources/pri_form.png" alt="System with one paraboloid reflector" width="700px"/><figcaption>System with one paraboloid reflector</figcaption></figure>
+
+
+If adding the reflector did not succeed a message will be logged in the console with information about what went wrong.
+
+<figure><img src="README_Resources/emptyFieldErr.png" alt="text saying: Error - Empty field at Focus xyz" width="600px"/></figure>
+
+
+## Plotting
+Now to see the reflector we just defined we need to plot it. 
+By clicking on the three dots on the right side of the element widget the element options menu will open. The first option is plot. 
+
+<figure><img src="README_Resources/options_plot.png" width="300px"/><figcaption>Selecting Plot from the element options menu</figcaption></figure>
+
+
+This will show the plot on the plot screen.
+
+<figure><img src="README_Resources/pri_plot.png" width="600px"/><figcaption>Plot of the paraboloid reflector</figcaption></figure>
+
+
+## Defining Ray-Trace Frames
+Now we will define a ray trace frame. We can do that from the menu bar by selecting Ray-trace > Make frame > Tube.
+
+Again, a form will show up. We fill it with the next parameters.
+
+
+<figure><img src="README_Resources/frame_form.png" alt="Selecting make tube frame" /><figcaption>Frame parameter form</figcaption></figure>
+
+When the frame is added a widget will show up in the *Frames* tab of the workspace. By clicking *plot* in the frame options menu we are presented with a form that allows us to choose the plane in which we want to plot this frame. If we select x and y we get the following plot. 
+
+<figure><img src="README_Resources/frame_plot.png" width="800px" /><figcaption>Plot of the frame in xy plane</figcaption></figure>
+
+Now we've defined a frame on the xy plane with z-coordinate 0. We can translate this frame upwards by selecting the transform option from the frame options as shown below
+
+<figure><img src="README_Resources/frame_transrform.png" width="400px"/><figcaption>Frame transformation form</figcaption></figure>
+
+
+Now if we plot the frame again from the side (e.g. in the xz plane). After zooming in with the zoom tool in the plot, we can see that all of the points have z coordinate 3000.
+
[…]

[…]
