+++ src/common/InterfaceReflector.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <vector>
 #include <array>
-#define _USE_MATH_DEFINES
+#define M_PI 3.14159265358979323846
 #include <cmath>
 
 #include "Utils.h"
