### 2023-06-08 PyPO-dev/PyPO#131 @arend95
With approximation, we meant the treatment of electromagnetic fields as vectorial entities, of which the polarisation state can change due to propagation through systems containing curved surfaces.

I have studied the documentation of Prysm, and I think it is a very nice software. I am truly sorry if the paragraph in the paper misrepresented things in this way and I totally agree that this paragraph needs to be sharpened and revised for the JOSS paper.

From what I understood in the docs of Prysm, there are two fundamental differences (and one big overlap) in how Prysm and PyPO operate. In the following discussion, I will stick to using Prysm as the example because it is mentioned that Prysm is a superset of POPPY and PROPER, and I thus assume (and after having a look at the docs of both these packages) that Prysm, POPPY and PROPER work in a similar fashion, physics-wise.

Difference #1: polarisation
It seems that Prysm does plane-to-plane propagation using the Rayleigh-Sommerfeld integral. This integral is the Green's function based solution to the scalar Helmholtz equation on a plane. The solution proper is a convolution product of the Green's function with the scalar electric field distribution on the source plane. PyPO, on the other hand, uses the equivalent surface current approach, which is a solution to the vectorial Helmholtz equation using induced surface currents on scattering objects. This method involves using the boundary conditions of Maxwells equations at a perfectly reflecting surface to calculate the electric and magnetic surface currents that are induced. The solution to this vectorial integral is also in the form of a Green's function (the overlap between Prysm and PyPO!) convolution, but the amplitudes and phases of the resulting EM-field components are affected by these surface currents because of the convolution of the Green's function with the induced surface currents.

It seems that in Prysm, polarisation is only taken into account when calculating transmission/reflection coefficients for thin films (I claim this by a search for "polarization" in the Prysm docs). Here, the polarisation is used in the context of the local plane of incidence, judging from the two options "p" and "s". This is  the "approximation" we mentioned in the paragraph, the capability of PyPO to calculate (induced) polarisation.

Thus, PyPO is capable of propagating polarised EM fields (see the tutorials in the docs where we actually calculate this in the form of the cross-polar efficiency of an optical system as a demonstration of this claim). This is interesting for applications where (induced) polarisation is of concern, such as design of radio telescopes and satellite dishes. See. This induced polarisation is brought forth by the curvature of optical elements which raises the next difference.

Difference #2: propagation and design
In Prysm, propagation is done it seems by defining an aperture distribution on a plane. Then, aberrations are added to the model by expanding them into (orthogonal) polynomials acting on the shape of the wavefront. Then, using the Rayleigh-Sommerfeld integral with Fourier approach (with or without cosine obliquity), the source distribution is propagated through a system consisting of a number of planes, which I presume terminates at a plane as well. PyPO, however, propagates field distributions to (off-axis) quadric or planar surfaces, calculates the induced surface currents at these surfaces using the boundary conditions of Maxwell's equations at perfectly reflecting interfaces, and uses these calculated currents to calculate the EM-field incident on the next reflector, which can be curved as well. The placement of these curved reflectors in 3D space and the ability to visualise these curved surfaces also seems to be a difference between PyPO and Prysm. Therefore, I think that another differnce between PyPO and Prysm is in how curved surfaces are handled and how this influences the design process.

In summary, I think Prysm and PyPO are two different beasts, designed for different purposes, with different design principles.

Would it be acceptable for you if we replace the quoted paragraph in the JOSS paper by the following:

_Commercial software for GO and PO calculations and analysis, such as OpticStudio (Zemax) and GRASP (TICRA), has already been developed. Of these, GRASP also employs the method of equivalent surface currents. Open-source optical simulation software packages such as POPPy (Perrin & others, 2012) and Prysm (Dube, 2019) are also available. These open-source software packages use the Rayleigh-Sommerfeld integral equation, which is a solution to the scalar Helmholtz equation, and thus do not model polarisation of the propagated fields. PyPO aims to contribute to this open-source optical simulation ecosystem by offering a software package working on the principle of the equivalent surface currents, which allows for a fully vectorial propagation between multiple planar or (off-axis) quadric reflector surfaces. This is interesting in cases where polarisation is of concern, for example for the design of radio telescopes and satellite dishes, where receivers are highly sensitive to polarisation._

That point about the ray-trace focus finding being inaccurate was due to a default value set to 1e-2, a typo on our side. Now, in the current tutorial (not the GUI one I believe, this still needs to be fixed), the focus finding is more accurate (defaults to 1e-12).
