### 2023-06-08 d9d9249
Delete CONTRIBUTING.md
