+++ CONTRIBUTING.md
@@ -1,2 +1,0 @@
-People wishing to contribute are more than welcome to do so. Contributions can be made by opening a pull request to the main repository from your own fork of PyPO, containing your contributions. 
-These will then be reviewed and considered for addition to PyPO. Issues and bugs can be reported in the PyPO [issue tracker](https://github.com/PyPO-dev/PyPO/issues). 
