### 2023-06-08 PyPO-dev/PyPO#129 @arend95
The documentations have been updated to include more tutorials.
Specifically, we added an extra set that showcases the fundamentals of using PyPO, such as input dictionary types for public API methods, gridding options for reflectors, etc.
Also, the first GUI tutorial now contains a first section on how to launch the GUI.

There was another set of tutorials in the form of interactive jupyter-notebooks. These were a bit more on using PyPO in scripts and the functionalities.
These tutorials were already present when we first submitted PyPO, but in a different repository on the organisation page, so maybe hard/impossible to find without further instructions, which were indeed lacking in the original docs.
We have now put them in the main repository, under the tutorials/ directory. Also, the docs now include these tutorials as statically rendered HTML.

We have now made a separate API reference, containing all public methods for interacting with PyPO.
In addition, we have added a full software documentation that documents the entire backend of PyPO.

A separate page has been made describing how contributions should be carried out using the fork-and-branch workflow. This page also contains guidelines on how to comment code so that documentation can be generated from it using doxygen, for example.

Is the current state of the docs acceptable?
