### 2023-06-08 PyPO-dev/PyPO#130 @arend95
Thanks for the comments! We will get to this ASAP. The second author @MrGafaji  has a busy week this week, but we plan to work on it this weekend. We will keep you updated on this issue!
