### 2023-06-09 PyPO-dev/PyPO#129 @brandondube
Thank you for greatly enhancing the documentation!  A few residual comments:

On https://pypo-dev.github.io/PyPO/basictut1.html

- Typo: `PyPO uses the [CMake]() build system to compile` (missing/incomplete/broken link)
- `which is part of the standard packages on Python (maybe mention PEP number?).` I don't think the parenthetical (note-to-self) should be in the released version of the documentation
- The section that discusses memory management seems somewhat inconsistent; it mentions that Python handles all (de-)allocation, but at the end says `the bottom layer deletes all internally allocated memory structures.`, so both layers do their own memory management?
- Typo: `visualisation if optical systems and results,`
- I am not sure this page is really a tutorial?  The user is not shown any code, is not asked to write any code, and is not shown any steps/processes/etc that would normally be a tutorial.  I think this page would be more of an "Explanation" in [The Documentation System](https://documentation.divio.com/)

On https://pypo-dev.github.io/PyPO/basictut2.html

- I am not sure this is a tutorial either?  It shows how to import the object and how to use one method, related to logging.  But it doesn't show me how to actually _do_ anything with the system

On https://pypo-dev.github.io/PyPO/basictut3.html

- Missing link: `A Numpy array of length 2, this parameter sets the number of cells along the gridding axes. Read more here.` (last 3 words)
- `A special case arises when 'a' = 'b'. In this case, the paraboloid is a so-called [surface of revolution](https://en.wikipedia.org/wiki/Surface_of_revolution), and is symmetric around the central symmetry axis. In this case, the paraboloid has two focii: one at a finite distance from the vertex (the bottom of the 'bowl' of the paraboloid) along the axis of symmetry and one at infinity.` this sentence is not really correct/in keeping with the norms of optics. A focus is a definite point, a focus at infinity does not really make any sense.
- `Hyperboloid Geometries` -- elsewhere in the docs, 'quadric' is used to describe the type of surface the software is able to model.  A hyperbola is not a quadratic, but spheres, parabolas, hyperbolas, and ellipses are all conics.  I think the documentation (and paper) should say conic where they currently say quadric.
- This page also feels more like an API reference, and not a tutorial.  I am not able to do anything with these types of surfaces by reading this page; actually, I do not even know how to (in code) describe one of these surfaces with the software!  There is no line of code showing how to use anything, just the meaning of the parameters (an API reference)

The refrain that these are not really tutorials applies to the "gridding" and "some more" pages, as well
