### 2023-06-09 PyPO-dev/PyPO#129 @arend95
Thanks a lot for the detailed comments, especially the link to the documentation system! Wish I knew about that earlier...

I have moved all these "basic tutorials" to a new explanations page, as they are indeed more like explanations than tutorials. Also, I agree with and have implemented almost all other residual comments.

The only point I do not agree with is the renaming of the surfaces from quadric to conic. A conic, in my understanding, is the curve traced out by intersecting a plane with a cone. The curve itself is described as the locus of a quadratic equation in two variables (a quadratic curve), and hence the set of points describing the curve is inherently 1D, in the sense that you can continuously map the conic onto the line of real numbers.

A quadric surface, on the other hand, is defined as the locus of a quadratic equation in three variables.
Since the hyperboloid equation consists of terms of at most degree 2, I think it is still a quadric surface.

Please see [this link](https://www.win.tue.nl/~sterk/Bouwkunde/hoofdstuk3.pdf). On page 13 (69 on the printed document), the general equation for a quadric surface is given, which I believe also encapsulates the equation for a hyperboloid of two sheets (the upper sheet is used in PyPO). Also, on page 21 (77 in print), they discuss the hyperboloid as a quadric surface, so I do believe that "quadric surface" is an appropriate term here.
