### 2023-06-09 PyPO-dev/PyPO#128 @arend95
Thank you for the comment!
I have now included a prerequisites subsection to the "install from source" section where I explicitly mention that cmake has to be pip-installed before building from source, irrespective of operating system.

This afternoon, we will merge with the main branch again. When you resume the review, can you let us know if the installation instructions work? If it all works, I will close this issue.
