### 2023-06-09 PyPO-dev/PyPO#131 @arend95
Ah, thank you for the detailed explanation. I will then update the paper with the edited paragraph and take your suggestion into account. I will let you know when its merged into main.

I will also have a look at the missing GUI images and fix that. The Jupyter notebooks do have the improved focus finder as well so there it can be checked.

And I am sorry about not mentioning Prysm's polarisation module. I could not find it in the docs so thought it was not there. But seeing now it is experimental, I fully understand that it is not included yet.

Could you let  me know if you agree with the updated paragraph in the newly compiled JOSS paper as a more accurate state of the field? Then I will close this issue afterwards.
