+++ tutorials/Gui/Tutorial_Gui_3.md
@@ -1,6 +1,6 @@
-# PyPO GUI Tutorial: performing physical optics propagations.
+﻿# PyPO GUI Tutorial: performing physical optics propagations.
 
-In this tutorial we will perform of a physical optics propagation on the same optical system  
+In this tutorial we will perform a physical optics propagation through the same optical system as the previous tutorial.
 
 ## Setup
 We start by loading our saved system from the previous tutorial. From the *Systems* menu we select *Load system*
@@ -13,12 +13,17 @@ We remove the paraboloid from the element options in the workspace:
 
 <img src="README_Resources/ResourcesT3/removeElem.png" alt="" width="400px"/>
 
-And we add a new Paraboloid
+and we add a new paraboloid:
 
 <img src="README_Resources/ResourcesT3/newPar.png" alt="" width="400px"/>
 
+We do the same to the tilted plane. We remove the old one, increase the gridsizes from 101 by 101 to 301 by 301 and reduce the radius from 10 mm to 5 mm.
 
-Next we define a circular plane (by giving it uv limits rather than xy) in the upper focus of the ellipsoid. 
+If we do not do this, the beam pattern we obtain on the paraboloid is undersampled coming from the tilted plane and we would see nonsensical results.
+
+We have also replaced the ellipsoidal mirror with a new one, with gridsizes of 279 by 279 (just like the one in the Jupyter notebook tutorial 3) an have reduced the radius to u = 10 mm. It is important to check the *Flip Normal Vectors* box, because we are illuminating the ellipsoid from below.
+
+Next we define a circular plane (by giving it uv limits rather than xy) in the upper focus of the ellipsoid. This will serve as the focal plane of our ellipsoid on which we define the initial beam pattern. 
 
 <img src="README_Resources/ResourcesT3/pl_foc.png" alt="" width="400px"/>
 
@@ -32,16 +37,16 @@ After clicking *Add beam* we can see that a field and a current have been added 
 
 <img src="README_Resources/ResourcesT3/beamCreated.png" alt="" width="300px"/>
 
-Now we can translate our *plane_focus* to position at (0, 0, 100).
+Now we can translate our *plane_focus* with 100 mm along the z-axis.
 
 <img src="README_Resources/ResourcesT3/transFocPl.png" alt="" width="300px"/>
 
 ## Propagation
-Now we can propagate our beam through the system. From the *Physical-optics* menu we select *Propagate beam > To surface*. The we do the following propagations  
+Now we can propagate our beam through the system. From the *Physical-optics* menu we select *Propagate beam > To surface*. Then we perform the following propagations  
 <img src="README_Resources/ResourcesT3/poProp1.png" alt="" width="300px"/><img src="README_Resources/ResourcesT3/poProp2.png" alt="" width="300px"/>
 <img src="README_Resources/ResourcesT3/poProp3.png" alt="" width="300px"/>
 
-We set, for the first two propagations, the 'mode' parameter to 'JM'. This means we only store the calculated JM currents on the target surface. If we specify 'EH', such as for the last propagation, we only save the illuminating field on the target surface. If we want both, we specify 'mode' as 'JMEH'. Another option, 'FF' for far-field, will be explained in more detail below. The last option, 'EHP', stores the reflected field and corresponding Poynting vectors. With this option it is possible to do a combined ray-trace and PO approach. This will be introduced in a later tutorial.
[…]
