+++ tests/test_BindUtils.py
@@ -1,0 +1,91 @@
+import unittest
+import numpy as np
+
+import matplotlib.pyplot as pt
+
+import PyPO.BindUtils as BU 
+import PyPO.Structs as PS
+
+
+class Test_BindUtils(unittest.TestCase):
+
+    def test_sfieldConv(self):
+        raise NotImplementedError()
+    
+    def test_arrC1ToObj(self):
+        raise NotImplementedError()
+    
+    def test_c2BundleToObj(self):
+        raise NotImplementedError()
+    
+    def test_c4BundleToObj(self):
+        raise NotImplementedError()
+    
+    def test_c2rBundleToObj(self):
+        raise NotImplementedError()
+    
+    def test_allocate_arrC1(self):
+        raise NotImplementedError()
+    
+    def test_allocate_c2Bundle(self):
+        raise NotImplementedError()
+    
+    def test_allocate_c4Bundle(self):
+        raise NotImplementedError()
+    
+    def test_allocate_c2rBundle(self):
+        raise NotImplementedError()
+    
+    def test_allfill_reflparams(self):
+        raise NotImplementedError()
+    
+    def test_allocate_reflcontainer(self):
+        raise NotImplementedError()
+    
+    def test_allocate_cframe(self):
+        raise NotImplementedError()
+    
+    def test_allfill_cframe(self):
+        raise NotImplementedError()
+    
+    def test_allfill_c2Bundle(self):
+        raise NotImplementedError()
+    
+    def test_allfill_RTDict(self):
+        raise NotImplementedError()
+    
+    def test_allfill_GRTDict(self):
+        raise NotImplementedError()
+    
+    def test_allfill_GPODict(self):
+        raise NotImplementedError()
+    
+    def test_allfill_SGPODict(self):
+        raise NotImplementedError()
+    
+    def test_allfill_mat4D(self):
+        raise NotImplementedError()
+    
+    def test_creflToObj(self):
+        raise NotImplementedError()
+    
+    def test_frameToObj(self):
+        raise NotImplementedError()
+    
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+if __name__ == "__main__":
+    unittest.main()
