### 2023-06-10 27b2487
Added function in TestTemplates to get full system
