+++ tests/TestUtils/TestTemplates.py
@@ -1,4 +1,5 @@
 import numpy as np
+import PyPO.System as System
 
 ##
 # @file
@@ -67,7 +68,7 @@ paraboloid_foc_xy = {
         "name"      : "testParaboloid_foc_xy",
         "pmode"     : "focus",
         "vertex"    : np.array([0, 0, 0]),
-        "focus"     : np.array([0, 0, 1]),
+        "focus_1"   : np.array([0, 0, 1]),
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
@@ -78,7 +79,7 @@ paraboloid_foc_uv = {
         "name"      : "testParaboloid_foc_uv",
         "pmode"     : "focus",
         "vertex"    : np.array([0, 0, 0]),
-        "focus"     : np.array([0, 0, 1]),
+        "focus_1"   : np.array([0, 0, 1]),
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
@@ -90,7 +91,7 @@ hyperboloid_man_xy = {
         "name"      : "testHyperboloid_man_xy",
         "pmode"     : "manual",
         "coeffs"    : np.array([1, 1, 1]),
-        "ecc"       : 0.5,
+        "ecc"       : 1.1,
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
@@ -101,7 +102,7 @@ hyperboloid_man_uv = {
         "name"      : "testHyperboloid_man_uv",
         "pmode"     : "manual",
         "coeffs"    : np.array([1, 1, 1]),
-        "ecc"       : 0.5,
+        "ecc"       : 1.1,
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
@@ -111,8 +112,9 @@ hyperboloid_man_uv = {
 hyperboloid_foc_xy = {
         "name"      : "testHyperboloid_foc_xy",
         "pmode"     : "focus",
-        "focus_1"   : np.array([0, 0, -1]),
-        "focus_2"   : np.array([0, 0, 1]),
+        "focus_1"   : np.array([0, 0, 1]),
+        "focus_2"   : np.array([0, 0, -1]),
+        "ecc"       : 1.1,
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
@@ -122,8 +124,9 @@ hyperboloid_foc_xy = {
 hyperboloid_foc_uv = {
         "name"      : "testHyperboloid_foc_uv",
         "pmode"     : "focus",
-        "focus_1"   : np.array([0, 0, -1]),
-        "focus_2"   : np.array([0, 0, 1]),
+        "focus_1"   : np.array([0, 0, 1]),
+        "focus_2"   : np.array([0, 0, -1]),
+        "ecc"       : 1.1,
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
@@ -159,8 +162,9 @@ ellipsoid_z_foc_xy = {
         "name"      : "testEllipsoid_z_foc_xy",
         "pmode"     : "focus",
         "orient"    : "z",
-        "focus_1"   : np.array([0, 0, -1]),
-        "focus_2"   : np.array([0, 0, 1]),
+        "focus_1"   : np.array([0, 0, 1]),
+        "focus_2"   : np.array([0, 0, -1]),
[…]

[…]
