+++ tests/test_SystemBeams.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import ctypes
 
-import TestUtils.TestTemplates as TestTemplates
+from . import TestTemplates
 
 import PyPO.BindBeam as beamlibs
 import PyPO.PyPOTypes as pypotypes
@@ -32,23 +32,10 @@ class Test_SystemBeams(unittest.TestCase):
         return flag
 
     def setUp(self):
-        self.s = System(verbose=False)
+        self.s = TestTemplates.getSystemWithReflectors()
 
     def test_loadBeamlib(self):
         lib = beamlibs.loadBeamlib()
-        self.assertEqual(type(lib), ctypes.CDLL)
-    
-    def test_loadTransflib(self):
-        lib = transflibs.loadTransflib()
-        self.assertEqual(type(lib), ctypes.CDLL)
-    
-
-    def test_loadCPUlib(self):
-        lib = cpulibs.loadCPUlib()
-        self.assertEqual(type(lib), ctypes.CDLL)
-
-    def test_loadGPUlib(self):
-        lib = gpulibs.loadGPUlib()
         self.assertEqual(type(lib), ctypes.CDLL)
 
     def test_createRTframe(self):
@@ -60,62 +47,50 @@ class Test_SystemBeams(unittest.TestCase):
         self.assertEqual(type(self.s.frames["testGaussRTframe"]), pypotypes.frame)
 
     def test_createGauss(self):
-        self.s.addPlane(TestTemplates.plane_xy)
         self.s.createGaussian(TestTemplates.GPOfield, "testPlane_xy")
         self.assertEqual(type(self.s.fields["testGaussField"]), pypotypes.fields)
         self.assertEqual(type(self.s.currents["testGaussField"]), pypotypes.currents)
 
-        self.s.addPlane(TestTemplates.plane_uv)
         self.s.createGaussian(TestTemplates.GPOfield, "testPlane_uv")
         self.assertEqual(type(self.s.fields["testGaussField"]), pypotypes.fields)
         self.assertEqual(type(self.s.currents["testGaussField"]), pypotypes.currents)
     
     def test_createScalarGauss(self):
-        self.s.addPlane(TestTemplates.plane_xy)
         self.s.createScalarGaussian(TestTemplates.GPOfield, "testPlane_xy")
         self.assertEqual(type(self.s.scalarfields["testGaussField"]), pypotypes.scalarfield)
 
-        self.s.addPlane(TestTemplates.plane_uv)
         self.s.createScalarGaussian(TestTemplates.GPOfield, "testPlane_uv")
         self.assertEqual(type(self.s.scalarfields["testGaussField"]), pypotypes.scalarfield)
 
     def test_createPointSource(self):
-        self.s.addPlane(TestTemplates.plane_xy)
         self.s.createPointSource(TestTemplates.PS_Ufield, "testPlane_xy")
         self.assertEqual(type(self.s.fields["testPS_UField"]), pypotypes.fields)
         self.assertEqual(type(self.s.currents["testPS_UField"]), pypotypes.currents)
 
-        self.s.addPlane(TestTemplates.plane_uv)
         self.s.createPointSource(TestTemplates.PS_Ufield, "testPlane_uv")
         self.assertEqual(type(self.s.fields["testPS_UField"]), pypotypes.fields)
         self.assertEqual(type(self.s.currents["testPS_UField"]), pypotypes.currents)
     
     def test_createScalarPointSource(self):
-        self.s.addPlane(TestTemplates.plane_xy)
         self.s.createPointSourceScalar(TestTemplates.PS_Ufield, "testPlane_xy")
[…]
