### 2023-06-10 689a2e7
Made all tests automated
