+++ tests/test_SystemGPU.py
@@ -131,6 +131,21 @@ class Test_SystemGPU(unittest.TestCase):
                     self.assertEqual(type(self.s.fields["test_EH"]), pypotypes.fields)
                     self.assertEqual(type(self.s.frames["test_P"]), pypotypes.frame)
 
+    def test_runPO_FF(self):
+        for i, plane in enumerate(TestTemplates.getPlaneList()):
+            if i == 2:
+                break
+
+            for i, source in enumerate(TestTemplates.getPOSourceList()):
+                if i == 0:
+                    self.s.createGaussian(source, plane["name"])
+                
+                elif i == 1:
+                    self.s.createPointSource(source, plane["name"])
+                    
+                runPODict = self._get_runPODictFF(source["name"], TestTemplates.getPlaneList()[-1]["name"],"test_EH")
+                self.s.runPO(runPODict)
+                self.assertEqual(type(self.s.fields["test_EH"]), pypotypes.fields)
 
     def test_runRT(self):
         if not self.GPU_flag:
@@ -206,6 +221,19 @@ class Test_SystemGPU(unittest.TestCase):
                 "mode"      : "JMEH"
                 }
 
+        return runPODict
+    
+    def _get_runPODictFF(self, source_current, target, name_EH):
+        runPODict = {
+                "t_name"    : target,
+                "s_current" : source_current,
+                "epsilon"   : 10,
+                "exp"       : "fwd",
+                "device"    : "CPU",
+                "name_EH"   : name_EH,
+                "mode"      : "FF"
+                }
+    
         return runPODict
     
     def _get_runPODictEHP(self, source_current, target, name_EH, name_P):
