+++ tests/test_Plotting.py
@@ -1,6 +1,5 @@
 import unittest
-from PyPO.System import System
-# import TestTemplates 
+import warnings
 try:
     from . import TestTemplates
 except ImportError: 
@@ -8,34 +7,98 @@ except ImportError:
 
 from numpy import ndarray 
 from matplotlib.figure import Figure
-from matplotlib.pyplot import Axes
+from matplotlib.pyplot import Axes, close
+from mpl_toolkits.mplot3d.axes3d import Axes3D
+from PyPO.System import System
 
+##
+# @file
+# This file contains tests for the plotting functionalities. It tests wether the plotting functions in PyPO.System  
 
 
 stm = TestTemplates.getSystemWithReflectors()
 
 class Test_SystemOps(unittest.TestCase):
 
-    # def test_plotBeamCut(self):
-    #     # stm.createGaussian(TestTemplates.GPOfield, 'testPlane_xy')
+    def test_plotBeamCut(self):
+        fig, ax = stm.plotBeamCut(TestTemplates.GPOfield['name'], 'Ex', center=False, align=False, ret=True)
 
-    #     # print(stm.fields,'!!!!!!!!')
+        self.assertEqual(type(fig), Figure)
+        self.assertEqual(type(ax), Axes)
 
-    #     fig, ax = stm.plotBeamCut(TestTemplates.GPOfield['name'], 'Ex', ret = True)
-
-    #     self.assertEqual(type(fig), Figure)
-    #     self.assertEqual(type(ax), Axes)
+        close('all')
 
     def test_plotBeam2D(self):
-
-        fig1, ax1 = stm.plotBeam2D(TestTemplates.GPOfield['name'], 'Ex', ret = True, amp_only=True)
-        fig2, ax2 = stm.plotBeam2D(TestTemplates.GPOfield['name'], 'Ex', ret = True)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            fig1, ax1 = stm.plotBeam2D(TestTemplates.GPOfield['name'], 'Ex', ret=True, amp_only=True)
+            fig2, ax2 = stm.plotBeam2D(TestTemplates.GPOfield['name'], 'Ex', ret=True)
 
         self.assertEqual(type(fig1), Figure)
         self.assertEqual(type(ax1), Axes)
         self.assertEqual(type(fig2), Figure)
         self.assertEqual(type(ax2), ndarray)
+
+        close('all')
+
+    def test_plotBeam3D(self):
+        fig1, ax1 = stm.plot3D(TestTemplates.paraboloid_man_uv['name'], ret=True)
+        self.assertEqual(type(fig1), Figure)
+        self.assertEqual(type(ax1), Axes3D)
+
+        for refl in (TestTemplates.getPlaneList() + 
+                     TestTemplates.getParaboloidList() + 
+                     TestTemplates.getHyperboloidList() +
+                     TestTemplates.getEllipsoidList()):
+            fig, ax = stm.plot3D(refl['name'], ret=True)
+
+            self.assertEqual(type(fig), Figure)
+            self.assertEqual(type(ax), Axes3D)
+
+            close('all')
+
+            
+    def test_plotSystem(self):
+        fig, ax = stm.plotSystem(ret=True)
+
+        self.assertEqual(type(fig), Figure)
+        self.assertEqual(type(ax), Axes3D)
+
+        emptySys = System(verbose=False)
+        figE, axE = emptySys.plotSystem(ret=True)
+
+        self.assertEqual(type(figE), Figure)
+        self.assertEqual(type(axE), Axes3D)
         
+
+        close('all')
+        
+        
+    def test_plotGroup(self):
+        stm.groupElements('testGroup', 
+                          TestTemplates.paraboloid_man_xy['name'],
[…]
