+++ tests/test_NumberingOfElements.py
@@ -47,95 +47,28 @@ class Test_SystemDictsAndAddElement(unittest.TestCase):
             self.s.createTubeFrame(TestTemplates.TubeRTframe)
             self.assertTrue(TestTemplates.TubeRTframe["name"] + f"_{i+1}" in self.s.frames)
         
-        for i in range(2):
             self.s.createGRTFrame(TestTemplates.GaussRTframe)
             self.assertTrue(TestTemplates.GaussRTframe["name"] + f"_{i+1}" in self.s.frames)
 
     def test_addPOFields(self):
-        validPlane = {
-            "name"      : "plane",
-            "gmode"     : "xy",
-            "lims_x"    : np.array([-1,1]),
-            "lims_y"    : np.array([-1,1]),
-            "gridsize"  : np.array([3, 3])
-            }
-        PSBeam = {
-            "name"          : 'PSBeam',
-            "lam"           : 1,
-            "E0"            : 1,
-            "phase"         : 0,
-            }
-        ScalarPSBeam = {
-            "name"          : 'SPSBeam',
-            "lam"           : 1,
-            "E0"            : 1,
-            "phase"         : 0,
-            "pol"           : np.array([0,0,1]),
-            }
-        UPSBeam = {
-            "name"          : 'UPSBeam',
-            "lam"           : 1,
-            "E0"            : 1,
-            "phase"         : 0,
-            "pol"           : np.array([0,0,1]),
-            }
-        ScalarUPSBeam = {
-            "name"          : 'SUPSBeam',
-            "lam"           : 1,
-            "E0"            : 1,
-            "phase"         : 0,
-            "pol"           : np.array([0,0,1]),
-            }
-        GaussBeam = {
-            "name"          : 'GaussBeam',
-            "lam"           : 1,
-            "w0x"           : 1,
-            "w0y"           : 1,
-            "n"             : 1,
-            "E0"            : 1,
-            "phase"         : 1,
-            "pol"           : np.array([0,0,1]),
-            }
-        ScalarGaussBeam = {
-            "name"          : 'SGaussBeam',
-            "lam"           : 1,
-            "w0x"           : 1,
-            "w0y"           : 1,
-            "n"             : 1,
-            "E0"            : 1,
-            "phase"         : 1,
-            "pol"           : np.array([0,0,1]),
-            }
-        self.s.addPlane(validPlane)
-        for func, validElem in [
-            (self.s.createPointSource, PSBeam), 
-            (self.s.createUniformSource, UPSBeam), 
-            (self.s.createGaussian, GaussBeam),
-            ]:
-            names = [validElem['name'] + '_1', validElem['name'] + '_2']
-            func(validElem, "plane")
-            func(validElem, "plane")
-            func(validElem, "plane")
-
-            for n in names:
-                self.assertTrue(n in self.s.fields)
-                self.assertTrue(n in self.s.currents)
+        for plane in TestTemplates.getPlaneList():
+            if plane["gmode"] == "AoE":
+                break
+        
+        for i in range(2):
+            self.s.createPointSource(TestTemplates.PS_Ufield, plane["name"])
+            self.assertTrue(TestTemplates.PS_Ufield["name"] + f"_{i+1}" in self.s.fields)
+            self.assertTrue(TestTemplates.PS_Ufield["name"] + f"_{i+1}" in self.s.currents)
             
+            self.s.createGaussian(TestTemplates.GPOfield, plane["name"])
+            self.assertTrue(TestTemplates.GPOfield["name"] + f"_{i+1}" in self.s.fields)
+            self.assertTrue(TestTemplates.GPOfield["name"] + f"_{i+1}" in self.s.currents)
+        
[…]
