+++ tests/test_SystemCopyAndRemove.py
@@ -1,159 +1,95 @@
 import unittest
-import numpy as np
-from PyPO.System import System as PyPOSystem
 from PyPO.Checks import InputReflError, InputRTError, InputPOError
 
+try:
+    from . import TestTemplates
+except ImportError:
+    import TestTemplates
 ##
 # @file
 # File containing tests for copy and remove methods for reflectors, fields, currents, frames and scalar fields.
 class Test_SystemCopyAndRemove(unittest.TestCase):
-    def setUp(self) -> None:
-        self.s = PyPOSystem(override=False, verbose=False)
-        
-        Parabola = {
-            "name"      : "par",
-            "pmode"     : "focus",
-            "gmode"     : "uv",
-            "vertex"    : np.zeros(3),
-            "focus_1"   : np.array([0,0,12e3]),
-            "lims_u"    : np.array([200,12.5e3]),
-            "lims_v"    : np.array([0, 360]),
-            "gridsize"  : np.array([1501,1501])
-            }
-
-        Hyperbola = {           
-            "name"      : "hyp",
-            "pmode"     : "focus",
-            "gmode"     : "uv",
-            "flip"      : True,
-            "focus_1"   : np.array([0,0,3.5e3]),
-            "focus_2"   : np.array([0,0,3.5e3 - 5606.286]),
-            "ecc"       : 1.08208248,
-            "lims_u"    : np.array([0,310]),
-            "lims_v"    : np.array([0,360]),
-            "gridsize"  : np.array([501,1501])
-            }
-
-        Ellipse = {
-            "name"      : "ell",
-            "pmode"     : "manual",
-            "gmode"     : "xy",
-            "coeffs"    : np.array([3199.769638 / 2, 3199.769638 / 2, 3689.3421 / 2]),
-            "flip"      : False,
-            "lims_x"    : np.array([1435, 1545]),
-            "lims_y"    : np.array([-200, 200]),
-            "gridsize"  : np.array([401, 401])
-            }
-
-        Plane = {
-            "name"      : "plane",
-            "gmode"     : "xy",
-            "lims_x"    : np.array([-0.1,0.1]),
-            "lims_y"    : np.array([-0.1,0.1]),
-            "gridsize"  : np.array([3, 3])
-            }
-
-        self.s.addParabola(Parabola)
-        self.s.addHyperbola(Hyperbola)
-        self.s.addEllipse(Ellipse)
-        self.s.addPlane(Plane)
-
-        RTDict = {
-            "name"          : "frame",
-            "nRays"         : 0,
-            "nRing"         : 0,
-            "angx0"         : 0,
-            "angy0"         : 0,
-            "x0"            : 4000,
-            "y0"            : 4000,
-            }
-
-        self.s.createTubeFrame(RTDict)
-        
-        PODict = {
-            "name"          : 'beam',
-            "lam"           : 1,
-            "E0"            : 1,
-            "phase"         : 0,
-            "pol"           : np.array([1,0,0]),
-            }
-
-        self.s.createPointSource(PODict, "plane")
-        self.s.createPointSourceScalar(PODict, "plane")
+    def setUp(self):
+        self.s = TestTemplates.getSystemWithReflectors()
+        self.s.setOverride(False) 
 
     def test_copyElement(self):
-        self.s.copyElement("par", "par_copy")
-        self.s.copyElement("hyp", "hyp_copy")
-        self.s.copyElement("ell", "ell_copy")
-        self.s.copyElement("plane", "plane_copy")
+        for plane in TestTemplates.getPlaneList():
+            self.s.copyElement(plane["name"], plane["name"] + "_copy")
+            self.assertTrue(plane["name"] + "_copy" in self.s.system)
 
[…]

[…]
