### 2023-06-12 PyPO-dev/PyPO#131 @brandondube
I believe the revised paragraph accurately captures the state of the field.  As a grammar thing, at the moment it is now `... surface currents ... fourier methods ... surface currents`.  You may consider re-ordering the sentences to colocate all of the surface current related sentences.
