### 2023-06-12 PyPO-dev/PyPO#130 @arend95
Thanks for the wait!

We have made all the tests automated as per the JOSS checklist. Also, tests such as the test_Bessel.py now do what they were supposed to do. The correct packaging of PyPO should have gotten rid of this particular code smell.
We also removed the parts , such as plotting code, that were implying manual testing in order to avoid confusion for people wanting to inspect the tests.

The instructions for running the testsuite can be found in the docs.
Instructions for generating a coverage overview, if desired, are included in the docs as well.
