### 2023-06-17 PyPO-dev/PyPO#150 @brandondube
Non-normative docstrings
This issue is part of my portion of the JOSS review, https://github.com/openjournals/joss-reviews/issues/5478

The revised documentation's contributing page shows how to document code, and elucifies that you are using something other than python docstrings to document functions.  This choice breaks completions for editors, jupyter kernels, etc, and will hamper users' ability to interact in the usual ways with your code when consuming it.  I would recommend documenting the python API using any of the standards for python docstrings; numpy style, google style, sphinx style, etc.

This is not mandatory, just a suggestion to help your software gain traction among python users
