### 2023-06-17 PyPO-dev/PyPO#130 @brandondube
Thank you for the updates.  A pain point, the documentation suggests: `./DevUtils.py -t`.  This is not normative for python (related to #150).  A python programmer would expect to find the instructions, e.g., `pytest .` or `python -m unittest ...` to invoke the unit test suite.  The documentation also relies on the devutils files having executable permissions for the user, and `/usr/bin/python3` being the interpreter that PyPO is installed to.  Please change this to at a minimum `python DevUtils.py -t`, so that the instructions are compatible with the full range of possible setups that `pip install .` can produce.

I ran the tests on my M2 macbook air for more than ten minutes, and ctrl+C'd them after waiting that long.  It looks like TestBessel has a dead spin or something in it,
```
# duration: 13m 13.44s
python DevUtils.py -c
Warning: you need to install "coverage_plugin" extra requirements to use this plugin. e.g. `pip install nose2[coverage_plugin]`
test_addLoggingLevel (tests.test_CustomLogger.TestCustomLogger) ... ok
test_CustomFormatter (tests.test_CustomLogger.TestCustomLogger) ... ok
test_CustomGUIFormatter (tests.test_CustomLogger.TestCustomLogger) ... ok
test_CustomGUILogger (tests.test_CustomLogger.TestCustomLogger) ... ok
test_CustomLogger (tests.test_CustomLogger.TestCustomLogger) ... ok
test_GUILogger (tests.test_CustomLogger.TestCustomLogger) ... ok
test_findConnectedSubsets (tests.test_MatUtils.Test_MatUtils) ... ok
test_plotBeam2D (tests.test_Plotting.Test_SystemOps) ... ok
test_plotBeam3D (tests.test_Plotting.Test_SystemOps) ... ok
test_plotBeamCut (tests.test_Plotting.Test_SystemOps) ... ok
test_plotGroup (tests.test_Plotting.Test_SystemOps) ... ok
test_plotRTframe (tests.test_Plotting.Test_SystemOps) ... ok
test_plotSystem (tests.test_Plotting.Test_SystemOps) ... ok
test_findRotation (tests.test_RayTraceUtils.Test_RayTraceUtils) ... ok
test_FocusFind_ellipse_x (tests.test_RayTraceUtils.Test_RayTraceUtils) ... ok
test_FocusFind_ellipse_z (tests.test_RayTraceUtils.Test_RayTraceUtils) ... ok
test_FocusFind_parabola (tests.test_RayTraceUtils.Test_RayTraceUtils) ... ok
test_getAnglesFromMatrix (tests.test_RayTraceUtils.Test_RayTraceUtils) ... ok
test_setCustomBeamPath (tests.test_SystemPaths.Test_SystemPaths) ... ok
test_setSavePath (tests.test_SystemPaths.Test_SystemPaths) ... ok
test_generateGridsEllipsoid (tests.test_SystemGrids.Test_SystemGrids) ... ok
test_generateGridsHyperboloid (tests.test_SystemGrids.Test_SystemGrids) ... ok
test_generateGridsParaboloid (tests.test_SystemGrids.Test_SystemGrids) ... ok
test_generateGridsPlane (tests.test_SystemGrids.Test_SystemGrids) ... ok
test_loadRefllib (tests.test_SystemGrids.Test_SystemGrids) ... ok
test_addGroup (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_addPOFields (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_addPOScalarFields (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_addReflector (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_addRTFrame (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_dictsExist (tests.test_SystemDictsAndAddElements.Test_SystemDictsAndAddElement) ... ok
test_mergeBeams_uv (tests.test_MergeBeams.Test_MergeBeams) ... ok
test_mergeBeams_xy (tests.test_MergeBeams.Test_MergeBeams) ... ok
test_fitGauss (tests.test_FitGauss.Test_FitGauss) ... ok
test_autoConverge (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_invalidRunHybridDict (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_invalidRunPODict (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_invalidRunRTDict (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_loadCPUlib (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_loadGPUlib (tests.test_SystemPO_RT.Test_SystemPO_RT) ... No GPU libraries found... Not testing GPU functionalities.
ok
test_runPO_EH (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runPO_EHP_Hybrid (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runPO_FF (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runPO_JM (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runPO_JMEH (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runPO_Scalar (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_runRT (tests.test_SystemPO_RT.Test_SystemPO_RT) ... ok
test_rotations_ellipse (tests.test_MatTransform.Test_MatTransform) ... ok
test_rotations_hyperbola (tests.test_MatTransform.Test_MatTransform) ... ok
test_rotations_parabola (tests.test_MatTransform.Test_MatTransform) ... ok
test_rotations_plane (tests.test_MatTransform.Test_MatTransform) ... ok
test_translations_ellipse (tests.test_MatTransform.Test_MatTransform) ... ok
test_translations_hyperbola (tests.test_MatTransform.Test_MatTransform) ... ok
test_translations_parabola (tests.test_MatTransform.Test_MatTransform) ... ok
test_translations_plane (tests.test_MatTransform.Test_MatTransform) ... ok
test_Bessel (tests.test_Bessel.Test_Bessel) ... ^C
Ran 56 tests in 787.694s

OK

Traceback (most recent call last):
  File "/Users/bdd/miniconda3/envs/tmp/bin/nose2", line 8, in <module>
    sys.exit(discover())
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/site-packages/nose2/main.py", line 339, in discover
    return main(*args, **kwargs)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/site-packages/nose2/main.py", line 87, in __init__
    super().__init__(**kw)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/main.py", line 101, in __init__
    self.runTests()
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/site-packages/nose2/main.py", line 302, in runTests
    self.result = runner.run(self.test)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/site-packages/nose2/runner.py", line 55, in run
    executor(test, result)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/site-packages/nose2/runner.py", line 43, in executor
    return suite(result)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 84, in __call__
    return self.run(*args, **kwds)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 122, in run
    test(result)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 84, in __call__
    return self.run(*args, **kwds)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 122, in run
    test(result)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 84, in __call__
    return self.run(*args, **kwds)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/suite.py", line 122, in run
    test(result)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/case.py", line 650, in __call__
    return self.run(*args, **kwds)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/case.py", line 591, in run
    self._callTestMethod(testMethod)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/unittest/case.py", line 549, in _callTestMethod
    method()
  File "/Users/bdd/src/PyPO/tests/test_Bessel.py", line 24, in test_Bessel
    e, h = self._calc_ff(lam)
  File "/Users/bdd/src/PyPO/tests/test_Bessel.py", line 83, in _calc_ff
    s.runPO(source_to_ff)
  File "/Users/bdd/src/PyPO/src/PyPO/System.py", line 1174, in runPO
    out = PyPO_CPUd(source, target, _runPODict)
  File "/Users/bdd/src/PyPO/src/PyPO/BindCPU.py", line 210, in PyPO_CPUd
    mgr.new_sthread(target=lib.propagateToFarField, args=args)
  File "/Users/bdd/src/PyPO/src/PyPO/Threadmgr.py", line 20, in new_sthread
    t.join(.1)
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/threading.py", line 1100, in join
    self._wait_for_tstate_lock(timeout=max(timeout, 0))
  File "/Users/bdd/miniconda3/envs/tmp/lib/python3.10/threading.py", line 1116, in _wait_for_tstate_lock
    if lock.acquire(block, timeout):
KeyboardInterrupt
```
The other tests took just a few moments to run up to then.  Could you check if the bessel unit tests are just some arm64 / big endian bug or something?
