### 2023-06-17 PyPO-dev/PyPO#129 @brandondube
Thank you; I looked over the written documentation, and it looks good.  I was not able to check that the GUI works as described in the docs.  The PySide2 dependency is wheel-only, and does not support darwin-arm64:

https://pypi.org/project/PySide2/#files

With regards to quadratic; my apologies, we come from different backgrounds.  I assumed quadric was a synonym or alternative spelling for "quadratic."  It would appear it is really a specific class of geometry, which all of the contemporary conics of optics are members of.  You may find users from an optics background are confused by the term as I have been, but it is technically correct as-is
