+++ src/GUI/WorkSpace.py
@@ -1,5 +1,5 @@
-from PySide2.QtWidgets import QWidget, QTabWidget, QScrollArea, QVBoxLayout, QSizePolicy
-from PySide2.QtCore import Qt
+from PySide6.QtWidgets import QWidget, QTabWidget, QScrollArea, QVBoxLayout, QSizePolicy
+from PySide6.QtCore import Qt
 from src.GUI.ElementWidget import ReflectorWidget, GroupWidget, FrameWidget, CurrentWidget, FieldsWidget, CurrentWidget, SFieldsWidget
 from src.GUI.Accordion import Accordion
 
