+++ src/GUI/Dialogs.py
@@ -1,5 +1,5 @@
-from PySide2.QtWidgets import QDialog, QGridLayout, QPushButton, QLabel
-from PySide2.QtCore import Qt
+from PySide6.QtWidgets import QDialog, QGridLayout, QPushButton, QLabel
+from PySide6.QtCore import Qt
 
 ##
 # @file Defines dialogs for the GUI.
