+++ src/GUI/ElementWidget.py
@@ -1,5 +1,5 @@
-from PySide2.QtWidgets import QLabel, QWidget, QHBoxLayout, QPushButton, QVBoxLayout
-from PySide2.QtCore import Qt
+from PySide6.QtWidgets import QLabel, QWidget, QHBoxLayout, QPushButton, QVBoxLayout
+from PySide6.QtCore import Qt
                                         
 from src.GUI.Dialogs import selfClosingDialog
 from src.GUI.utils import MyButton
