### 2023-06-17 0988e2a
Changed structure of tests
