+++ tests/TestTemplates.py
@@ -27,7 +27,7 @@ GaussRTframe =  {
 
 GPOfield =  {
         "name"      : "testGaussField",
-        "lam"       : 1,
+        "lam"       : 0.01,
         "w0x"       : 1,
         "w0y"       : 1,
         "n"         : 1,
@@ -51,7 +51,7 @@ paraboloid_man_xy = {
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 paraboloid_man_uv = {
@@ -61,7 +61,7 @@ paraboloid_man_uv = {
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 paraboloid_foc_xy = {
@@ -72,7 +72,7 @@ paraboloid_foc_xy = {
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 paraboloid_foc_uv = {
@@ -83,7 +83,7 @@ paraboloid_foc_uv = {
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 #### HYPERBOLOIDS ####
@@ -95,7 +95,7 @@ hyperboloid_man_xy = {
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 hyperboloid_man_uv = {
@@ -106,7 +106,7 @@ hyperboloid_man_uv = {
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 hyperboloid_foc_xy = {
@@ -118,7 +118,7 @@ hyperboloid_foc_xy = {
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 hyperboloid_foc_uv = {
@@ -130,7 +130,7 @@ hyperboloid_foc_uv = {
         "gmode"     : "uv",
         "lims_u"    : np.array([0, 1]),
         "lims_v"    : np.array([0, 360]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 #### ELLIPSOIDS ####
@@ -143,7 +143,7 @@ ellipsoid_z_man_xy = {
         "gmode"     : "xy",
         "lims_x"    : np.array([-1, 1]),
         "lims_y"    : np.array([-1, 1]),
-        "gridsize"  : np.array([3, 3])
+        "gridsize"  : np.array([13, 13])
         }
 
 ellipsoid_z_man_uv = {
@@ -155,7 +155,7 @@ ellipsoid_z_man_uv = {
[…]
