### 2023-06-18 01ee974
Moved manual tests to demos
