+++ src/PyPO/System.py
@@ -28,8 +28,6 @@ import PyPO.FitGauss as fgs
 
 import PyPO.WorldParam as world
 
-# Set PyPO absolute root path
-sysPath = Path(__file__).parents[2]
 logging.getLogger(__name__)
 
 ##
@@ -42,15 +40,15 @@ class System(object):
     customReflPath = os.getcwd()
     savePathSystems = os.getcwd()
 
-
-    ##
-    # Constructor. Initializes system state.
-    #
-    # @param redirect Redirect all print statements within system to given stdout.
-    # @param context Whether system is created in script or in GUI.
-    # @param verbose Enable system logger.
-    # @param override Allow overriding names.
     def __init__(self, redirect=None, context=None, verbose=True, override=True):
+        """!
+        Constructor. Initializes system state.
+        
+        @param redirect Redirect all print statements within system to given stdout.
+        @param context Whether system is created in script or in GUI.
+        @param verbose Enable system logger.
+        @param override Allow overriding names.
+        """
         self.num_ref = 0
         self.num_cam = 0
         self.nThreads_cpu = os.cpu_count()
@@ -61,17 +59,11 @@ class System(object):
         Config.setOverride(override)
 
         self.system = {}
-        
         self.frames = {}
-        
         self.fields = {}
-        
         self.currents = {}
-        
         self.scalarfields = {}
-
         self.groups = {}
-        
         self.assoc = {}
 
         self.EHcomplist = np.array(["Ex", "Ey", "Ez", "Hx", "Hy", "Hz"])
@@ -109,29 +101,33 @@ class System(object):
         if override == True:
             self.clog.warning("System override set to True.")
        
-    ##
-    # Destructor. Deletes any reference to the logger assigned to current system.
     def __del__(self):
+        """!
+        Destructor. Deletes any reference to the logger assigned to current system.
+        """
         if self.context != "G":
             self.clog.info("EXITING SYSTEM.")
             del self.clog_mgr
             del self.clog
     
-    ##
-    # @brief Obtain a reference to the custom logger used by system.
-    #
-    # This method can be called to obtain a reference to the logging object that PyPO uses internally.
-    # Can be convenient in cases one wants to log their own information in the layout of the PyPO logger.
-    #
-    # @ingroup public_api_sysio
-    #
-    # @returns clog Reference to system logger.
     def getSystemLogger(self):
+        """!
+        @brief Obtain a reference to the custom logger used by system.
+        
+        This method can be called to obtain a reference to the logging object that PyPO uses internally.
+        Can be convenient in cases one wants to log their own information in the layout of the PyPO logger.
+        
+        @ingroup public_api_sysio
+        
+        @returns clog Reference to system logger.
+        """
         return self.clog
 
-    ##
-    # Print system contents.
     def __str__(self):
+        """!
+        Print system contents.
+        """
+        
         s = "Contents of system:\n"
         s += f"Reflectors {len(self.system)}:\n"
         for key in self.system.keys():
@@ -153,32 +149,35 @@ class System(object):
             s += f"    {key}\n"
         return s
 
-    ##
-    # @brief Set path to folder containing custom beam patterns.
-    #
-    # Set the path to the directory where PyPO looks for custom beam patterns.
-    # When a custom beam pattern is imported using the readCustomBeam() method, PyPO will look in the specified path for the beam patterns. 
-    #
-    # @ingroup public_api_sysio
-    #
-    # @param path Path to custom beams.
-    # @param append Whether path is relative to current working directory or absolute from home.
     def setCustomBeamPath(self, path, append=False):
+        """!
+        Set path to folder containing custom beam patterns.
+        
[…]
