### 2023-06-19 PyPO-dev/PyPO#129 @arend95
Thank you for the observation on the arm64. We have changed the dependency on PySide2 to PySide6 - according to its [PyPi page](https://pypi.org/project/PySide6/#files), it has a universal2 wheel for macosx. From what I could find, this also includes arm64 support.

Could you please confirm if PySide6 installs correctly, and with that, if the GUI can be started?

And I agree that the term "quadric" could be confusing for people used to certain conventions from optics, even though it is a mathematically correct term. We have now included a few lines in the explanation on "Reflectors/elements in PyPO" where we state the common usage of "conics" in optics and why we choose to use the term "quadric". I hope that this clears up some potential confusion.
