### 2023-06-19 f93d4c8
Streamlined tests and fixed broken link in docs
