+++ requirements.txt
@@ -1,15 +1,0 @@
-numpy==1.24.3
-contourpy==1.0.7
-cycler==0.11.0
-fonttools==4.39.3
-importlib-resources==5.12.0
-kiwisolver==1.4.4
-matplotlib==3.7.1
-nose2==0.12.0
-packaging==23.1
-Pillow==9.5.0
-pyparsing==3.0.9
-python-dateutil==2.8.2
-scipy==1.10.1
-six==1.16.0
-zipp==3.15.0

[…]
