### 2023-06-19 PyPO-dev/PyPO#150 @arend95
Thank you for the suggestion. We have investigated the possibilities and have come to the conclusion that doxygen does support special commands in docstrings, with the help of a little extra formatting:
"""!
<docstring>
"""
This is supported by doxygen and its special commands, but comes at the cost of an extra exclamation mark after the first quote triplet which might not comply with a common style but at least adds the docstring to the __doc__ attribute of a function.
We have already implemented this in the entire src/PyPO/ directory (source directory), but not yet for the GUI source code. We will do this in due time.

The updated documenting style (using docstrings) is now also reflected in the documentation itself.

Could you please let us know if this satisfies your suggestion?
