### 2023-06-19 8138767
Docs in GUI converted to docstrings
