+++ src/GUI/ParameterForms/formData.py
@@ -5,39 +5,45 @@ from src.GUI.ParameterForms.InputDescription import inType, InputDescription
 # This file contains all forms used by the GUI for interaction with the user.
 # Because all functions return lists containing forms, the return will not be explicitly stated per form.
 
-##
-# Options for generating an element from an xy parametrisation.
+
 def xy_opts():
+    """!
+    Options for generating an element from an xy parametrisation.
+    """
     return [InputDescription(inType.vectorFloats, "lims_x", label="X limits", oArray=True, numFields=2),
             InputDescription(inType.vectorFloats, "lims_y", label="Y limits", oArray=True, numFields=2)]
 
-##
-# Options for generating an element from a uv parametrisation.
 def uv_opts():
+    """!
+    Options for generating an element from a uv parametrisation.
+    """
     return [InputDescription(inType.vectorFloats, "lims_u", label="U limits", oArray=True, numFields=2, toolTip = "U limits: inner and outer diameter."),
             InputDescription(inType.vectorFloats, "lims_v", label="V limits", prefill = True, hints=[0., 360.], oArray=True, numFields=2, toolTip = "V angle in degrees."),
             InputDescription(inType.vectorFloats, "gcenter", label="XY center", hints = [0.,0.], oArray=True, numFields=2, prefill = True),
             InputDescription(inType.vectorFloats, "ecc_uv", label="UV eccentricity", numFields=1, hints = [0.], prefill = True),
             InputDescription(inType.vectorFloats, "rot_uv", label="UV position angle", numFields=1, hints = [0.], prefill = True)]
 
-##
-# Options for generating a far-field element from an AoE (Azimuth-over-Elevation) parametrisation.
 def AoE_opts():
+    """!
+    Options for generating a far-field element from an AoE (Azimuth-over-Elevation) parametrisation.
+    """
     return [InputDescription(inType.vectorFloats, "lims_Az", label="Azimuth limits", oArray=True, numFields=2),
             InputDescription(inType.vectorFloats, "lims_El", label="Elevation limits", oArray=True, numFields=2)]
 
-##
-# Extra options for hyperboloids and ellipsoids for supplying quadric parameters and gridding options.
 def focus_opts_hyp_ell():
+    """!
+    Extra options for hyperboloids and ellipsoids for supplying quadric parameters and gridding options.
+    """
     return [
         InputDescription(inType.vectorFloats, "focus_1", label="Upper focus xyz", oArray=True, numFields=3),
             InputDescription(inType.vectorFloats, "focus_2", label="Lower focus xyz", oArray=True, numFields=3),
             InputDescription(inType.vectorFloats, "ecc", label="Eccentricity", numFields=1)
     ]
 
-##
-# Form for generating a paraboloid element.
 def makeParabolaInp():
+    """!
+    Form for generating a paraboloid element.
+    """
     return [
         InputDescription(inType.vectorStrings, "name", hints = ["Parabola"], prefill = True),
         InputDescription(inType.dynamicRadio, "pmode", label="Parameter mode", subDict={
@@ -53,9 +59,10 @@ def makeParabolaInp():
         })
     ]
 
-##
-# Form for generating a hyperboloid element.
 def makeHyperbolaInp():
+    """!
+    Form for generating a hyperboloid element.
+    """
     return [InputDescription(inType.vectorStrings, "name", hints = ["Hyperbola"], prefill = True),
             InputDescription(inType.dynamicRadio, "pmode", label="Parameter mode", subDict={
                 "focus"     : focus_opts_hyp_ell(),
@@ -68,9 +75,10 @@ def makeHyperbolaInp():
                 "uv" : uv_opts()
             })]
 
-##
-# Form for generating an ellipsoid element.
 def makeEllipseInp():
+    """!
[…]
