+++ src/PyPO/BindCPU.py
@@ -2,9 +2,9 @@ import ctypes
 import numpy as np
 import os
 import pathlib
-from PyPO.BindUtils import *
-from PyPO.Structs import *
-from PyPO.PyPOTypes import *
+
+import PyPO.BindUtils as BUtils
+import PyPO.Structs as PStructs
 import PyPO.Config as Config
 import PyPO.Threadmgr as TManager
 
@@ -29,50 +29,50 @@ def loadCPUlib():
         except:
             lib = ctypes.CDLL(os.path.join(path_cur, "libpypocpu.dylib"))
 
-    lib.propagateToGrid_JM.argtypes = [ctypes.POINTER(c2Bundle), reflparams, reflparams,
-                                        ctypes.POINTER(reflcontainer), ctypes.POINTER(reflcontainer),
-                                        ctypes.POINTER(c2Bundle),ctypes.c_double, ctypes.c_int,
+    lib.propagateToGrid_JM.argtypes = [ctypes.POINTER(PStructs.c2Bundle), PStructs.reflparams, PStructs.reflparams,
+                                        ctypes.POINTER(PStructs.reflcontainer), ctypes.POINTER(PStructs.reflcontainer),
+                                        ctypes.POINTER(PStructs.c2Bundle),ctypes.c_double, ctypes.c_int,
                                         ctypes.c_double, ctypes.c_double]
 
     lib.propagateToGrid_JM.restype = None
 
-    lib.propagateToGrid_EH.argtypes = [ctypes.POINTER(c2Bundle), reflparams, reflparams,
-                                        ctypes.POINTER(reflcontainer), ctypes.POINTER(reflcontainer),
-                                        ctypes.POINTER(c2Bundle),ctypes.c_double, ctypes.c_int,
+    lib.propagateToGrid_EH.argtypes = [ctypes.POINTER(PStructs.c2Bundle), PStructs.reflparams, PStructs.reflparams,
+                                        ctypes.POINTER(PStructs.reflcontainer), ctypes.POINTER(PStructs.reflcontainer),
+                                        ctypes.POINTER(PStructs.c2Bundle),ctypes.c_double, ctypes.c_int,
                                         ctypes.c_double, ctypes.c_double]
 
     lib.propagateToGrid_EH.restype = None
 
-    lib.propagateToGrid_JMEH.argtypes = [ctypes.POINTER(c4Bundle), reflparams, reflparams,
-                                        ctypes.POINTER(reflcontainer), ctypes.POINTER(reflcontainer),
-                                        ctypes.POINTER(c2Bundle),ctypes.c_double, ctypes.c_int,
+    lib.propagateToGrid_JMEH.argtypes = [ctypes.POINTER(PStructs.c4Bundle), PStructs.reflparams, PStructs.reflparams,
+                                        ctypes.POINTER(PStructs.reflcontainer), ctypes.POINTER(PStructs.reflcontainer),
+                                        ctypes.POINTER(PStructs.c2Bundle),ctypes.c_double, ctypes.c_int,
                                         ctypes.c_double, ctypes.c_double]
 
     lib.propagateToGrid_JMEH.restype = None
 
-    lib.propagateToGrid_EHP.argtypes = [ctypes.POINTER(c2rBundle), reflparams, reflparams,
-                                        ctypes.POINTER(reflcontainer), ctypes.POINTER(reflcontainer),
-                                        ctypes.POINTER(c2Bundle),ctypes.c_double, ctypes.c_int,
[…]

[…]
