### 2023-06-20 PyPO-dev/PyPO#150 @arend95
The GUI source code has now also been updated to contain the docstrings instead of the old style.
