### 2023-06-22 0ce32dd
Updated docs for correct pyqtdarktheme
