+++ src/GUI/MainWindow.py
@@ -45,6 +45,7 @@ class MainWidget(QWidget):
         super().__init__(parent)
         # Window settings
         self.setWindowTitle("PyPO")
+        self.currentFileName = ""
 
         # GridParameters
         self.grid = QGridLayout()
@@ -380,25 +381,47 @@ class MainWidget(QWidget):
     
     ### Functionalities: Systems 
 
-    def saveSystemForm(self):
+    def saveSystem(self):
         """!
         Opens a form that allows user to save the System.
         """
+        try:
+            if not self.currentFileName:
+                self.saveSystemAs()
+                return
+            self.stm.saveSystem(self.currentFileName)
+        except Exception as err:
+            print(err)
+            print_tb(err.__traceback__)
+            self.clog.error(err)
         
         # self.setForm(fData.saveSystemForm(), readAction=self.saveSystemAction, okText="Save System") ##TODO delete form
     
-    def saveSystemAction(self):
+    def saveSystemAs(self):
         """!
         Saves the current system state under the name given in form.
         """
+        
         try:
             diag = QFileDialog(self)
             diag.setFileMode(QFileDialog.FileMode.AnyFile)
             homedir = os.path.expanduser('~')
-            out, _ = diag.getSaveFileName(self, filter="*.pyposystem", dir = homedir)
-            str_l = out.rsplit(sep = os.sep, maxsplit = 1)
-            self.stm.setSavePathSystems(str_l[0])
-            self.stm.saveSystem(str_l[1])
+            filePath, _ = diag.getSaveFileName(self, filter="*.pyposystem", dir = homedir)
+            if not filePath:
+                return
+            pathFileList = filePath.rsplit(sep = os.sep, maxsplit = 1)
+            if pathFileList[1] == (self.currentFileName + ".pyposystem"):
+                pathFileList[1] = self.currentFileName
+            
+            self.currentFileName = pathFileList[1]
+            # print(pathFileList)
+            # print(self.currentFileName)
+            # sss = pathFileList.split('.')
+            # if len() == 3:
+            #     pathFileList = 
+
+            self.stm.setSavePathSystems(pathFileList[0])
+            self.saveSystem()
 
         except Exception as err:
             print(err)
@@ -423,13 +446,7 @@ class MainWidget(QWidget):
             print(err)
             print_tb(err.__traceback__)
             self.clog.error(err)
-    def loadSystemForm(self):
-        """!
-        Opens a form that allows user to load a saved System.
-        """
-        systemList = [os.path.split(x[0])[-1] for x in os.walk(self.stm.savePathSystems) if os.path.split(x[0])[-1] != "systems"]
-        self.setForm(fData.loadSystemForm(systemList), readAction=self.loadSystemAction, okText="Load System")
-    
+ 
     def loadSystem(self):
         """!
         Loads system selected in from form.
@@ -443,32 +460,16 @@ class MainWidget(QWidget):
             if diag.exec_():
                 
                 str_l = diag.selectedFiles()[0]
-                print(str_l)
 
                 str_l = str_l.rsplit(sep = os.sep, maxsplit = 1)
                 self.stm.setSavePathSystems(str_l[0])
                 self.stm.loadSystem(str_l[1].split(".")[0])
+                self.currentFileName = str_l[1].split(".")[0]
 
-
+                self.refreshWorkspaceSection()
                 
-
-            # out, _ = diag.getSaveFileName(self, filter="*.pyposystem", dir = homedir)
-            # str_l = out.rsplit(sep = os.sep, maxsplit = 1)
-            # self.stm.setSavePath(str_l[0])
-            # self.stm.saveSystem(str_l[1])
-
[…]
