### 2023-06-25 PyPO-dev/PyPO#150 @brandondube
Thank you.  One nit: your docs say:

```
Because docstrings cannot be used as file descriptions, we use a different style to describe files. Just as for C/C++/CUDA, file descriptions go after import statements and before the classes and/or methods in the file:
```

If the first line of a python file is a string, that is the docstring for the file.  For example, prysm:

https://github.com/brandondube/prysm/blob/master/prysm/__init__.py#L1

and

```
>>> prysm.__doc__
'prysm, a python optics module.'
```
