### 2023-06-25 PyPO-dev/PyPO#129 @brandondube
Thanks, all of that works well.  One mild nit -- on MacOS, Qt conforms to the OS norms, and the menu bar is used.  In, e.g. [this tutorial](https://pypo-dev.github.io/PyPO/Gui/Tutorial_Gui_1.html), you say

> From the elements menu select Add Reflector > Quadric surface.

My first thought on reading this was that the element menu was the set of four tabs beneath the logo (on macos), because the menubar is at the top of the screen / outside the actual application window.  You might consider adding a bit of clarification there

Otherwise, everything LGTM
