### 2023-06-25 PyPO-dev/PyPO#130 @brandondube
thanks!  They all pass with a `nose2` command now
