### 2023-06-26 ecb57f7
Added pyqtdarktheme as requirement in paper
