### 2023-06-28 c40bf36
Started unit testing cpp, mostly for dev reasons. Refactored a bit
