+++ GenerateDocs.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import traceback
+import argparse
 
 ##
 # @file
@@ -8,55 +9,67 @@ import traceback
 #
 # For this script to work properly, you should have installed the docs prerequisites.
 def GenerateDocs():
-        tut_path = "tutorials"
-        demo_path = "demos"
-
+    docPath = os.path.join(os.path.dirname(os.path.realpath(__file__)), "docs")
+    tutPath = os.path.join(os.path.dirname(os.path.realpath(__file__)), "tutorials")
+    demoPath = os.path.join(os.path.dirname(os.path.realpath(__file__)), "demos")
+    doxyPath = os.path.join(os.path.dirname(os.path.realpath(__file__)), "doxy")
+    
+    parser = argparse.ArgumentParser(description="options for generating docs")
+    parser.add_argument("-t", "--tutorials", help="generate docs from tutorials", action="store_true")
+    parser.add_argument("-g", "--guitutorials", help="generate GUI tutorial docs", action="store_true")
+    parser.add_argument("-d", "--demos", help="generate docs from demos", action="store_true")
+    args = parser.parse_args()
+    
+    try:
         try:
-            try:
-                shutil.rmtree("docs")
-            except Exception as err:
-                print(traceback.format_exc())
+            shutil.rmtree(docPath)
+        except Exception as err:
+            print(traceback.format_exc())
 
+        if args.tutorials:
             # Convert regular tutorials to html format for inclusion in the documentation.
-            for (dirpath, dirnames, filenames) in os.walk(tut_path):
-                dest_path = os.path.join("docs", "tutorials")
-                os.makedirs(dest_path)
+            for (dirpath, dirnames, filenames) in os.walk(tutPath):
+                destPath = os.path.join(docPath, "tutorials")
+                os.makedirs(destPath)
                 for file in filenames:
                     if file.split('.')[1] != "ipynb":
                         continue
 
-                    _path = os.path.join(tut_path, file)
-                    html_path = os.path.join(tut_path, f"{file.split('.')[0]}.html")
-                    html_dest_path = os.path.join(dest_path, f"{file.split('.')[0]}.html")
+                    _path = os.path.join(tutPath, file)
+                    htmlPath = os.path.join(tutPath, f"{file.split('.')[0]}.html")
+                    html_destPath = os.path.join(destPath, f"{file.split('.')[0]}.html")
 
                     os.system(f"jupyter nbconvert --to html --template lab --theme dark {_path}")
-                    os.rename(html_path, html_dest_path)
+                    os.rename(htmlPath, html_destPath)
                 
                 break
-            
+       
+        if args.demos:
             # Convert demos to html format for inclusion in the documentation.
-            for (dirpath, dirnames, filenames) in os.walk(demo_path):
-                dest_path = os.path.join("docs", "demos")
-                os.makedirs(dest_path)
+            for (dirpath, dirnames, filenames) in os.walk(demoPath):
+                destPath = os.path.join(docPath, "demos")
+                os.makedirs(destPath)
                 for file in filenames:
                     if file.split('.')[1] != "ipynb":
                         continue
 
-                    _path = os.path.join(demo_path, file)
-                    html_path = os.path.join(demo_path, f"{file.split('.')[0]}.html")
-                    html_dest_path = os.path.join(dest_path, f"{file.split('.')[0]}.html")
+                    _path = os.path.join(demoPath, file)
+                    htmlPath = os.path.join(demoPath, f"{file.split('.')[0]}.html")
+                    html_destPath = os.path.join(destPath, f"{file.split('.')[0]}.html")
 
                     os.system(f"jupyter nbconvert --to html --template lab --theme dark {_path}")
-                    os.rename(html_path, html_dest_path)
[…]
