+++ src/include/tests/test_GUtils.cpp
@@ -1,0 +1,253 @@
+#include <gtest/gtest.h>
+#include "../GUtils.h"
+#include <array>
+#include <cmath>
+#include <cuComplex.h>
+#define M_PI 3.14159265358979323846
+
+class GUtilsTest : public ::testing::Test {
+    protected:
+    
+    cuFloatComplex z0 = {1, 1};
+    cuFloatComplex z0c = {1, -1};
+    cuFloatComplex z0m = {-1, 1};
+    cuFloatComplex c0 = {0, 0};
+    cuFloatComplex j = {0, 1};
+    
+    float mu = 1.;
+    float err = 1e-6;
+
+    float ad [3]= {1, 1, 1};
+    cuFloatComplex cad [3]= {z0, z0, z0};
+
+    float xax[3] = {1., 0, 0};
+    float yax[3] = {0, 1., 0};
+    float zax[3] = {0, 0, 1.};
+    float zax_fl[3] = {0, 0, -1.};
+
+    float rot[3] = {0., 0., 90 * M_PI / 180};
+    float crot[3] = {};
+
+    float out_d;
+    cuFloatComplex cout_d;
+
+    float ad_out[3];
+    cuFloatComplex cad_out[3];
+
+    float md_out[3][3];
+    float eye[3][3] = {};
+    float eye4[16] = {};
+
+    void SetUp() override {
+        this->eye[0][0] = 1.;
+        this->eye[1][1] = 1.;
+        this->eye[2][2] = 1.;
+        
+        this->eye4[0] = 1.;
+        this->eye4[5] = 1.;
+        this->eye4[10] = 1.;
+        this->eye4[15] = 1.;
+    }
+    
+};
+
+bool equal(const cuFloatComplex &c1, const cuFloatComplex &c2) {
+        if(c1.x == c2.x && c1.y == c2.y)
+            return true;
+        else
+            return false;
+}
+
+TEST_F(GUtilsTest, TestDotProducts) {
+    dot(ad, ad, out_d); // Check real-real dot
+    EXPECT_EQ(out_d, 3.);
+    
+    dot(cad, cad, cout_d);
+    EXPECT_EQ(cout_d.x, 6.);
+    EXPECT_EQ(cout_d.y, 0.);
+
+    dot(cad, ad, cout_d);
+    EXPECT_EQ(cout_d.x, 3.);
+    EXPECT_EQ(cout_d.y, -3.);
+    
+    dot(ad, cad, cout_d);
+    EXPECT_EQ(cout_d.x, 3.);
+    EXPECT_EQ(cout_d.y, 3.);
+}
+
+TEST_F(GUtilsTest, TestExternalProducts) {
+    ext(ad, ad, ad_out); // Check parallel cross
+    EXPECT_EQ(ad_out[0], 0.);
+    EXPECT_EQ(ad_out[1], 0.);
+    EXPECT_EQ(ad_out[2], 0.);
+
+    ext(xax, yax, ad_out); // Check x out y = z
+    EXPECT_EQ(ad_out[0], zax[0]);
+    EXPECT_EQ(ad_out[1], zax[1]);
+    EXPECT_EQ(ad_out[2], zax[2]);
+    
+    ext(yax, xax, ad_out); // Check y out x = -z
+    EXPECT_EQ(ad_out[0], zax[0]);
+    EXPECT_EQ(ad_out[1], zax[1]);
[…]
