+++ src/GUI/MainWindow.py
@@ -16,6 +16,7 @@ from src.GUI.utils import inType
 import src.GUI.ParameterForms.formData as fData
 from src.GUI.PlotScreen import PlotScreen
 from src.GUI.WorkSpace import Workspace
+from src.GUI.Dialogs import UnsavedChangesDialog
 from src.GUI.SubprocessManager import SubprocessManager, copySystem
 
 from PyPO.CustomLogger import CustomGUILogger
@@ -46,6 +47,8 @@ class MainWidget(QWidget):
         # Window settings
         self.setWindowTitle("PyPO")
         self.currentFileName = ""
+        self.unsavedChanges = False
+        self.darkTheme  = True
 
         # GridParameters
         self.grid = QGridLayout()
@@ -73,12 +76,7 @@ class MainWidget(QWidget):
         self._mkPlotScreen()
         self.setLayout(self.grid)
         
-        self.event_stop = Event()
-
-        # self.threadpool = QThreadPool()
         self.subprocessManager = SubprocessManager(self)
-
-        self.frameDict = {}
 
     ### GUI setup functions
     
@@ -114,8 +112,11 @@ class MainWidget(QWidget):
             self.WorkSpace.setParent(None)
         # rebuild 
         logo = QLabel()
-        pixmap = QPixmap('src/GUI/resources/logo.png')
-        pixmap = pixmap.scaledToWidth(250)
+        imgUrl = 'src/GUI/resources/logo.png'
+        if self.darkTheme:
+            imgUrl = 'src/GUI/resources/logo_dark_bg.png'
+        pixmap = QPixmap(imgUrl)
+        pixmap = pixmap.scaledToWidth(300)
         logo.setPixmap(pixmap)
         logo.resize(300, 150)
         self.WorkSpace = Workspace()
@@ -269,6 +270,7 @@ class MainWidget(QWidget):
             
             self.stm.copyElement(copyDict["name"], copyDict["name_copy"])
             self.addReflectorWidget(copyDict["name_copy"])
+            self.unsavedChanges = True
 
 
         except Exception as err:
@@ -292,6 +294,7 @@ class MainWidget(QWidget):
             
             self.stm.copyGroup(copyDict["name"], copyDict["name_copy"])
             self.addGroupWidget(copyDict["name_copy"])
+            self.unsavedChanges = True
         except Exception as err:
             print(err)
             print_tb(err.__traceback__)
@@ -380,6 +383,11 @@ class MainWidget(QWidget):
             self.clog.error(err)
     
     ### Functionalities: Systems 
+    def saveBeforeClosing(self):
+        if self.unsavedChanges:
+            return UnsavedChangesDialog(self, self.saveSystem).exec_()
+        return True
+                
 
     def saveSystem(self):
         """!
@@ -387,9 +395,10 @@ class MainWidget(QWidget):
         """
         try:
             if not self.currentFileName:
-                self.saveSystemAs()
-                return
+                return self.saveSystemAs()
             self.stm.saveSystem(self.currentFileName)
+            self.unsavedChanges = False
+            return True
         except Exception as err:
             print(err)
             print_tb(err.__traceback__)
@@ -408,50 +417,31 @@ class MainWidget(QWidget):
             homedir = os.path.expanduser('~')
             filePath, _ = diag.getSaveFileName(self, filter="*.pyposystem", dir = homedir)
             if not filePath:
-                return
+                return False
             pathFileList = filePath.rsplit(sep = os.sep, maxsplit = 1)
             if pathFileList[1] == (self.currentFileName + ".pyposystem"):
                 pathFileList[1] = self.currentFileName
             
             self.currentFileName = pathFileList[1]
-            # print(pathFileList)
-            # print(self.currentFileName)
-            # sss = pathFileList.split('.')
[…]
