### 2023-06-28 d31a07e
Now enums in high level functions. Started implementing lenses
