+++ GuiApp.py
@@ -1,4 +1,8 @@
-# from __future__ import print_function
+"""!
+@file
+Main GUI entry point. Run this file from the terminal and watch the magic happen.
+"""
+
 import sys
 from PySide6.QtWidgets import QApplication
 from src.GUI.MainWindow import PyPOMainWindow
