+++ src/GUI/PlotScreen.py
@@ -1,10 +1,13 @@
+"""!
+@file 
+Defines a matplotlib plot widget that can be placed in the GUI.
+"""
+
 from PySide6.QtWidgets import  QWidget, QVBoxLayout
 from matplotlib import use
 from matplotlib.figure import Figure
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg, NavigationToolbar2QT as NavigationToolbar
 
-##
-# @file Defines a matplotlib plot widget that can be placed in the GUI.
 use('Qt5Agg')
 class MplCanvas(FigureCanvasQTAgg):
     def __init__(self, fig):
