+++ tests/test_SystemPO_RT.py
@@ -1,3 +1,8 @@
+"""!
+@file
+Tests for checking if CPU and GPU operations in PyPO are correct
+"""
+
 import unittest
 import numpy as np
 import ctypes
@@ -16,10 +21,7 @@ import PyPO.Templates as pytemp
 
 from PyPO.System import System
 from PyPO.Checks import RunPOError, RunRTError, HybridPropError, check_runPODict, check_hybridDict, check_runRTDict
-
-##
-# @file
-# Tests for checking if CPU and GPU operations in PyPO are correct
+from PyPO.Enums import FieldComponents, CurrentComponents
 
 class Test_SystemPO_RT(unittest.TestCase):
     def setUp(self):
@@ -85,9 +87,9 @@ class Test_SystemPO_RT(unittest.TestCase):
             self.assertEqual(type(self.s.currents["test_JM"]), pypotypes.currents)
             self.assertEqual(type(self.s.fields["test_EH"]), pypotypes.fields)
     
-    @params(("CPU", True, System.runPO, System.runHybridPropagation, None, np.ones(3)), 
-            ("GPU", True, System.runPO, System.runHybridPropagation, "Ex", None), 
-            ("CPU", False, System.runGUIPO, System.hybridGUIPropagation, "Ex", None))
+    @params(("CPU", True, System.runPO, System.runHybridPropagation, FieldComponents.NONE, np.ones(3)), 
+            ("GPU", True, System.runPO, System.runHybridPropagation, FieldComponents.Ex, None), 
+            ("CPU", False, System.runGUIPO, System.hybridGUIPropagation, FieldComponents.Ex, None))
     def test_runPO_EHP_Hybrid(self, dev, interp, funcPO, funcHybrid, comp, start):
         if dev == "GPU" and not self.GPU_flag:
             return
@@ -279,7 +281,7 @@ class Test_SystemPO_RT(unittest.TestCase):
 
         return runPODict
     
-    def _get_runPODictHybrid(self, fr_in, field_in, target, fr_out, field_out, interp=False, comp=None, start=None):
+    def _get_runPODictHybrid(self, fr_in, field_in, target, fr_out, field_out, interp=False, comp=FieldComponents.NONE, start=None):
         runPODict = {
                 "fr_in"     : fr_in,
                 "t_name"    : target,
