### 2023-08-12 #5478 @arend95
Hi @danielskatz,

The release tag for the archived version is v1.0.1 and the corresponding DOI on Zenodo is 10.5281/zenodo.8241427
I checked the metadata and made the specified corrections. I hope it is now good to go!
