+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @book{Balanis:1989,
-  author    = {Balanis, C. A.},
+  author    = {Balanis, Constantine A.},
   title     = {Advanced Engineering Electromagnetics},
   isbn      = {0-471-62194-3},
   publisher = {John Wiley \& Sons},
@@ -7,7 +7,7 @@
 }
 
 @article{Taniguchi:2022,
-  author    = {Taniguchi, A. and others},
+  author    = {Taniguchi, Akio and Bakx, Tom J. L. C. and Baselmans, Jochem J. A. and Huiting, Robert and Karatsu, Kenichi and Llombart, Nuria and Rybak, Matus and Takekoshi, Tatsuya and Tamura, Yoichi and Akamatsu, Hiroki and Brackenhoff, Stefanie and Bueno, Juan and Buijtendorp, Bruno T. and Dabironezare, Shahab O. and Doing, Anne-Kee and Fujii, Yasunori and Fujita, Kazuyuki and Gouwerok, Matthijs and Hähnle, Sebastian and Ishida, Tsuyoshi and Ishii, Shun and Kawabe, Ryohei and Kitayama, Tetsu and Kohno, Kotaro and Kouchi, Akira and Maekawa, Jun and Matsuda, Keiichi and Murugesan, Vignesh and Nakatsubo, Shunichi and Oshima, Tai and Pascual Laguna, Alejandro and Thoen, David J. and van der Werf, Paul P. and Yates, Stephen J. C. and Endo, Akira},
   title     = {DESHIMA 2.0: Development of an Integrated Superconducting Spectrometer for Science-Grade Astronomical Observations},
   journal   = {Journal of Low Temperature Physics},
   year      = {2022},
@@ -18,7 +18,7 @@
 }
 
 @inproceedings{ASTE,
-    author = {H. Ezawa and others},
+    author = {Ezawa, Hajime and Kawabe, Ryohei and Kohno, Kotaro and Yamamoto, Satoshi},
     title = {{The Atacama Submillimeter Telescope Experiment (ASTE)}},
     volume = {5489},
     booktitle = {Ground-based Telescopes},
@@ -59,7 +59,7 @@
 }
 
 @Article{Hunter:2007,
-  Author    = {Hunter, J. D.},
+  Author    = {Hunter, John D.},
   Title     = {Matplotlib: A 2D graphics environment},
   Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
@@ -75,7 +75,7 @@
 
 @Article{         Harris:2020,
  title         = {Array programming with {NumPy}},
- author        = {Charles R. Harris and others},
+ author        = {Harris, Charles R. and Millman, K. Jarrod and van der Walt, Stéfan J. and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J. and Kern, Robert and Picus, Matti and Hoyer, Stephan and van Kerkwijk, Marten H. and Brett, Matthew and Haldane, Allan and del Río, Jaime Fernández and Wiebe, Mark and Peterson, Pearu and Gérard-Marchant, Pierre and Sheppard, Kevin and Reddy, Tyler and Weckesser, Warren and Abbasi, Hameer and Gohlke, Christoph and Oliphant, Travis E.},
  year          = {2020},
  month         = sep,
  journal       = {Nature},
@@ -88,7 +88,7 @@
 }
 
 @ARTICLE{Virtanen:2020,
-  author  = {Virtanen, Pauli and others},
[…]
