### 2023-08-12 #5478 @danielskatz
@arend95 - I've suggested a bunch of fairly small changes in https://github.com/PyPO-dev/PyPO/pull/165 - please merge this, or let me know what you disagree with, then we can proceed.
