### 2023-08-12 119186b
Missing Oxford comma
Added final Oxford comma between matplotlib and scipy
