### 2023-08-12 077776c
Removed spurious vertical bar in paper.md
