### 2023-08-12 #5478 @arend95
@danielskatz - I found another missing Oxford comma in the paper.md file and went ahead and fixed it. I hope that the paper.md file is good to go now.
