+++ paper/paper.md
@@ -52,3 +52,22 @@ The GUI carries dependencies on PySide6, pyqtdarktheme, and attrs.
 Development of `PyPO` started with the need for alignment strategies for the wideband sub-mm spectrometer DESHIMA 2.0 [@Taniguchi:2022]. 
 A software package capable of efficient GO and PO calulations through optical systems consisting of off-axis quadric surfaces was necessary for calculating the configuration of the corrective optics. 
 Currently, `PyPO` is also being used in simulations of measured beam patterns of DESHIMA 2.0 at the ASTE [@ASTE] telescope for the analysis of instrument and optical system performance.
+
+Commercial software for GO/PO calculations and analysis, such as OpticStudio (Zemax) and GRASP (TICRA), has already been developed.
+Open-source optical simulation software packages such as POPPy (Perrin & others, 2012) and Prysm (Dube, 2019) are also available. 
+These open-source software packages use Fourier methods to solve the Rayleigh-Sommerfeld integral equation.
+`PyPO` aims to contribute to this open-source optical simulation ecosystem by offering a software package working on the principle of the equivalent surface currents.
+
+# Availability
+`PyPO` can be found on [Github](https://github.com/PyPO-dev/PyPO) and is available for Linux, MacOS, and Windows.
+Software documentation and instructions regarding installation, contributing and issue tracking can be found in the [documentation](https://pypo-dev.github.io/PyPO/).
+The package comes with several Jupyter Notebook tutorials illustrating the workflow and features, and can be used as building blocks for new reflector systems.
+Several tutorials for using the GUI features are also included and can be found in the software documentation.
+
+# Acknowledgements
+We thank Shahab Oddin Dabironezare for useful discussions and proofreading the manuscript.
+This work is supported by the European Union (ERC Consolidator Grant No. 101043486 TIFUUN). 
+Views and opinions expressed are however those of the authors only and do not necessarily reflect those of the European Union or the European Research Council Executive Agency. 
+Neither the European Union nor the granting authority can be held responsible for them.
+
+# References
