### 2023-08-12 #5478 @arend95
@danielskatz - I have fixed the missing parts of the paper. Sorry again, I am currently working on a tablet and the pen seems to live a life of its own sometimes. I checked the paper and it seems complete again.
