+++ paper/paper.md
@@ -59,7 +59,7 @@ These open-source software packages use Fourier methods to solve the Rayleigh-So
 `PyPO` aims to contribute to this open-source optical simulation ecosystem by offering a software package working on the principle of the equivalent surface currents.
 
 # Availability
-`PyPO` can be found on [Github](https://github.com/PyPO-dev/PyPO) and is available for Linux, MacOS, and Windows.
+`PyPO` can be found on [GitHub](https://github.com/PyPO-dev/PyPO) and is available for Linux, MacOS, and Windows.
 Software documentation and instructions regarding installation, contributing and issue tracking can be found in the [documentation](https://pypo-dev.github.io/PyPO/).
 The package comes with several Jupyter Notebook tutorials illustrating the workflow and features, and can be used as building blocks for new reflector systems.
 Several tutorials for using the GUI features are also included and can be found in the software documentation.
