### 2023-08-12 #5478 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
