### 2023-08-12 c7c1398
more small bib changes
