### 2023-08-12 #5478 @danielskatz
@arend95 - a few more small changes in the bib that I missed before - sorry this is taking so many iterations...

https://github.com/PyPO-dev/PyPO/pull/166
